Estimate contrasts

Estimate contrasts

# S3 method for stanreg
estimate_contrasts(
  model,
  levels = NULL,
  fixed = NULL,
  modulate = NULL,
  transform = "none",
  length = 10,
  standardize = TRUE,
  standardize_robust = FALSE,
  centrality = "median",
  ci = 0.95,
  ci_method = "hdi",
  test = c("pd", "rope"),
  rope_range = "default",
  rope_ci = 1,
  ...
)

Arguments

model
A Bayesian model.

levels
A character vector or formula specifying the names of the predictors over which to estimate means or contrasts.

fixed
A character vector indicating the names of the predictors to be "fixed" (i.e., maintained), so that the estimation is made at these values.

modulate
A character vector indicating the names of a numeric variable along which the means or the contrasts will be estimated. Adjust its length using length.

transform
Can be "none" (default for contrasts), "response" (default for means), "mu", "unlink", "log". "none" will leave the values on scale of the linear predictors. "response" will transform them on scale of the response variable. Thus for a logistic model, "none" will give estimations expressed in log-odds (probabilities on logit scale) and "response" in terms of probabilities.

length
Length of the spread numeric variables.

standardize
If TRUE, adds standardized differences or coefficients.

standardize_robust
Robust standardization through MAD (Median Absolute Deviation, a robust estimate of SD) instead of regular SD.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

ci
Credible Interval (CI) level. Default to 0.89 (89%). See ci for further details.

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi), "ETI" (see eti) or "SI" (see si).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests. For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

rope_range
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

...
Arguments passed to or from other methods.

Value

A data frame of estimated contrasts.

Examples

library(modelbased)

data <- mtcars
data$cyl <- as.factor(data$cyl)
data$am <- as.factor(data$am)

# \donttest{
if (require("rstanarm")) {

  model <- stan_glm(mpg ~ cyl * am, data = data, refresh=0)
  estimate_contrasts(model)
  estimate_contrasts(model, fixed = "am")

  model <- stan_glm(mpg ~ cyl * wt, data = data, refresh=0)
  estimate_contrasts(model)
  estimate_contrasts(model, fixed = "wt")
  estimate_contrasts(model, modulate = "wt", length = 4)
  estimate_contrasts(model, levels = "wt", length = 4)

  model <- stan_glm(Sepal.Width ~ Species + Petal.Width + Petal.Length, data = iris, refresh=0)
  estimate_contrasts(model, fixed = "Petal.Width", modulate = "Petal.Length", test = "bf")
}

#> Loading required package: rstanarm

#> Loading required package: Rcpp

#> rstanarm (Version 2.19.3, packaged: 2020-02-11 05:16:41 UTC)

#> - Do not expect the default priors to remain the same in future rstanarm versions.

#> Thus, R scripts should specify priors explicitly, even if they are just the defaults.

#> - For execution on a local, multicore CPU with excess RAM we recommend calling

#> options(mc.cores = parallel::detectCores())

#> - bayesplot theme set to bayesplot::theme_default()

#>    * Does _not_ affect other ggplot2 plots

#>    * See ?bayesplot_theme_set for details on theme setting

#> NOTE: Results may be misleading due to involvement in interactions

#> NOTE: Results may be misleading due to involvement in interactions

#> Computation of Bayes factors: sampling priors, please wait...

#> Loading required namespace: logspline

#> Level1     |     Level2 | Petal.Width | Petal.Length | Difference |       95% CI |    BF | Difference (std.)
#> ------------------------------------------------------------------------------------------------------------
#> setosa     | versicolor |        1.20 |         1.00 |       1.64 | [1.31, 1.99] | > 999 |              3.76
#> setosa     |  virginica |        1.20 |         1.00 |       2.01 | [1.53, 2.51] | > 999 |              4.62
#> versicolor |  virginica |        1.20 |         1.00 |       0.37 | [0.14, 0.57] |  7.88 |              0.85
#> setosa     | versicolor |        1.20 |         1.66 |       1.64 | [1.31, 1.99] | > 999 |              3.76
#> setosa     |  virginica |        1.20 |         1.66 |       2.01 | [1.53, 2.51] | > 999 |              4.62
#> versicolor |  virginica |        1.20 |         1.66 |       0.37 | [0.14, 0.57] |  7.88 |              0.85
#> setosa     | versicolor |        1.20 |         2.31 |       1.64 | [1.31, 1.99] | > 999 |              3.76
#> setosa     |  virginica |        1.20 |         2.31 |       2.01 | [1.53, 2.51] | > 999 |              4.62
#> versicolor |  virginica |        1.20 |         2.31 |       0.37 | [0.14, 0.57] |  7.88 |              0.85
#> setosa     | versicolor |        1.20 |         2.97 |       1.64 | [1.31, 1.99] | > 999 |              3.76
#> setosa     |  virginica |        1.20 |         2.97 |       2.01 | [1.53, 2.51] | > 999 |              4.62
#> versicolor |  virginica |        1.20 |         2.97 |       0.37 | [0.14, 0.57] |  7.88 |              0.85
#> setosa     | versicolor |        1.20 |         3.62 |       1.64 | [1.31, 1.99] | > 999 |              3.76
#> setosa     |  virginica |        1.20 |         3.62 |       2.01 | [1.53, 2.51] | > 999 |              4.62
#> versicolor |  virginica |        1.20 |         3.62 |       0.37 | [0.14, 0.57] |  7.88 |              0.85
#> setosa     | versicolor |        1.20 |         4.28 |       1.64 | [1.31, 1.99] | > 999 |              3.76
#> setosa     |  virginica |        1.20 |         4.28 |       2.01 | [1.53, 2.51] | > 999 |              4.62
#> versicolor |  virginica |        1.20 |         4.28 |       0.37 | [0.14, 0.57] |  7.88 |              0.85
#> setosa     | versicolor |        1.20 |         4.93 |       1.64 | [1.31, 1.99] | > 999 |              3.76
#> setosa     |  virginica |        1.20 |         4.93 |       2.01 | [1.53, 2.51] | > 999 |              4.62
#> versicolor |  virginica |        1.20 |         4.93 |       0.37 | [0.14, 0.57] |  7.88 |              0.85
#> setosa     | versicolor |        1.20 |         5.59 |       1.64 | [1.31, 1.99] | > 999 |              3.76
#> setosa     |  virginica |        1.20 |         5.59 |       2.01 | [1.53, 2.51] | > 999 |              4.62
#> versicolor |  virginica |        1.20 |         5.59 |       0.37 | [0.14, 0.57] |  7.88 |              0.85
#> setosa     | versicolor |        1.20 |         6.24 |       1.64 | [1.31, 1.99] | > 999 |              3.76
#> setosa     |  virginica |        1.20 |         6.24 |       2.01 | [1.53, 2.51] | > 999 |              4.62
#> versicolor |  virginica |        1.20 |         6.24 |       0.37 | [0.14, 0.57] |  7.88 |              0.85
#> setosa     | versicolor |        1.20 |         6.90 |       1.64 | [1.31, 1.99] | > 999 |              3.76
#> setosa     |  virginica |        1.20 |         6.90 |       2.01 | [1.53, 2.51] | > 999 |              4.62
#> versicolor |  virginica |        1.20 |         6.90 |       0.37 | [0.14, 0.57] |  7.88 |              0.85

if (require("brms")) {
  model <- brm(mpg ~ cyl * am, data = data, refresh=0)
  estimate_contrasts(model)
}

#> Loading required package: brms

#> Loading 'brms' package (version 2.13.0). Useful instructions
#> can be found by typing help('brms'). A more detailed introduction
#> to the package is available through vignette('brms_overview').

#> 
#> Attaching package: ‘brms’

#> The following objects are masked from ‘package:rstanarm’:
#> 
#>     dirichlet, exponential, get_y, lasso, ngrps

#> The following object is masked from ‘package:lme4’:
#> 
#>     ngrps

#> The following object is masked from ‘package:stats’:
#> 
#>     ar

#> Compiling the C++ model

#> Start sampling

#> Warning: Method 'parse_bf' is deprecated. Please use 'brmsterms' instead.

#> Warning: Method 'parse_bf' is deprecated. Please use 'brmsterms' instead.

#> Warning: Argument 'exact_match' is deprecated. Please use 'fixed' instead.

#> Level1 | Level2 | Difference |           95% CI |     pd | % in ROPE | Difference (std.)
#> ----------------------------------------------------------------------------------------
#> 4 0    |    6 0 |       3.79 | [ -0.77,   8.64] | 94.58% |     0.95% |              0.63
#> 4 0    |    8 0 |       7.83 | [  3.54,  11.74] | 99.98% |        0% |              1.30
#> 4 0    |    4 1 |      -5.18 | [ -9.48,  -1.13] | 99.20% |     0.18% |             -0.86
#> 4 0    |    6 1 |       2.30 | [ -2.74,   7.31] | 81.23% |     1.88% |              0.38
#> 4 0    |    8 1 |       7.53 | [  1.62,  12.89] | 99.22% |        0% |              1.25
#> 6 0    |    8 0 |       4.05 | [  0.57,   7.80] | 98.55% |     0.25% |              0.67
#> 6 0    |    4 1 |      -8.97 | [-12.71,  -4.99] |   100% |        0% |             -1.49
#> 6 0    |    6 1 |      -1.45 | [ -6.01,   3.63] | 74.17% |     2.97% |             -0.24
#> 6 0    |    8 1 |       3.73 | [ -1.46,   9.19] | 91.75% |     1.20% |              0.62
#> 8 0    |    4 1 |     -13.01 | [-15.95, -10.18] |   100% |        0% |             -2.16
#> 8 0    |    6 1 |      -5.51 | [ -9.67,  -1.60] | 99.48% |     0.18% |             -0.91
#> 8 0    |    8 1 |      -0.32 | [ -4.94,   4.46] | 55.07% |     3.48% |             -0.05
#> 4 1    |    6 1 |       7.46 | [  3.41,  11.89] | 99.90% |     0.03% |              1.24
#> 4 1    |    8 1 |      12.68 | [  7.53,  17.46] |   100% |        0% |              2.10
#> 6 1    |    8 1 |       5.21 | [ -0.73,  10.63] | 96.30% |     0.33% |              0.86

# }