Estimate marginal means

Estimate marginal means

# S3 method for lm
estimate_means(
  model,
  levels = NULL,
  fixed = NULL,
  modulate = NULL,
  transform = "response",
  length = 10,
  ci = 0.95,
  ...
)

Arguments

model
A Bayesian model.

levels
A character vector or formula specifying the names of the predictors over which to estimate means or contrasts.

fixed
A character vector indicating the names of the predictors to be "fixed" (i.e., maintained), so that the estimation is made at these values.

modulate
A character vector indicating the names of a numeric variable along which the means or the contrasts will be estimated. Adjust its length using length.

transform
Can be "none" (default for contrasts), "response" (default for means), "mu", "unlink", "log". "none" will leave the values on scale of the linear predictors. "response" will transform them on scale of the response variable. Thus for a logistic model, "none" will give estimations expressed in log-odds (probabilities on logit scale) and "response" in terms of probabilities.

length
Length of the spread numeric variables.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

...
Arguments passed to or from other methods.

Value

A data frame of estimated marginal means.

Examples

library(modelbased)

model <- lm(Petal.Length ~ Sepal.Width * Species, data = iris)

estimate_means(model)

#> NOTE: Results may be misleading due to involvement in interactions

estimate_means(model, fixed = "Sepal.Width")
estimate_means(model, levels = c("Species", "Sepal.Width"), length = 2)
estimate_means(model, levels = "Species=c('versicolor', 'setosa')")

#> NOTE: Results may be misleading due to involvement in interactions

estimate_means(model, levels = "Sepal.Width=c(2, 4)")

#> NOTE: Results may be misleading due to involvement in interactions

estimate_means(model, levels = c("Species", "Sepal.Width=0"))
estimate_means(model, modulate = "Sepal.Width", length = 5)
estimate_means(model, modulate = "Sepal.Width=c(2, 4)")
# \donttest{
if (require("lme4")) {
  data <- iris
  data$Petal.Length_factor <- ifelse(data$Petal.Length < 4.2, "A", "B")

  model <- lmer(Petal.Length ~ Sepal.Width + Species + (1 | Petal.Length_factor), data = data)
  estimate_means(model)
  estimate_means(model, modulate = "Sepal.Width", length = 3)
}# }