Generates predictions from models

estimate_link is a shortcut to estimate_response with data = "grid". estimate_response would be used in the context of generating actual predictions for the existing or new data, whereas estimate_link is more relevant in the context of visualisation and plotting. There are many control parameters that are not listed here but can be used, such as the arguments from visualisation_matrix (used when data = "grid") and from insight::get_predicted() (the function to compute predictions used internally).

estimate_relation(
  model,
  data = "grid",
  ci = 0.95,
  keep_iterations = FALSE,
  ...
)

estimate_link(
  model,
  data = "grid",
  predict = "link",
  ci = 0.95,
  keep_iterations = FALSE,
  ...
)

estimate_prediction(
  model,
  data = NULL,
  ci = 0.95,
  keep_iterations = FALSE,
  ...
)

estimate_response(model, data = NULL, ci = 0.95, keep_iterations = FALSE, ...)

Arguments

model
A statistical model.

data
A data frame with model's predictors to estimate the response. If NULL, the model's data is used. If "grid", the model matrix is obtained (through visualisation_matrix).

ci
The interval level (default 0.95, i.e., 95% CI).

keep_iterations
Only relevant for Bayesian models or simulated models. If TRUE, will keep all prediction iterations (draws). You can reshape them by running bayestestR::reshape_iterations().

...
You can add all the additional control arguments from visualisation_matrix (used when data = "grid") and insight::get_predicted().

predict
Can be "response" (default) or "link". The former predicts the the outcome per se, while the latter predicts the link function (i.e., the regression "line"), equivalent to estimating the fit. In other words, estimate_response(model, predict="link") is equivalent to estimate_link(model).

Value

A dataframe of predicted values.

Examples

library(modelbased)

# Linear Models
model <- lm(mpg ~ wt, data = mtcars)
estimate_response(model)

#>                        wt Predicted       SE    CI_low  CI_high
#> Mazda RX4           2.620 23.282611 3.111080 16.928937 29.63628
#> Mazda RX4 Wag       2.875 21.919770 3.099021 15.590725 28.24882
#> Datsun 710          2.320 24.885952 3.133524 18.486442 31.28546
#> Hornet 4 Drive      3.215 20.102650 3.093108 13.785680 26.41962
#> Hornet Sportabout   3.440 18.900144 3.095614 12.578057 25.22223
#> Valiant             3.460 18.793255 3.096084 12.470207 25.11630
#> Duster 360          3.570 18.205363 3.099389 11.875565 24.53516
#> Merc 240D           3.190 20.236262 3.093145 13.919216 26.55331
#> Merc 230            3.150 20.450041 3.093336 14.132605 26.76748
#> Merc 280            3.440 18.900144 3.095614 12.578057 25.22223
#> Merc 280C           3.440 18.900144 3.095614 12.578057 25.22223
#> Merc 450SE          4.070 15.533127 3.129637  9.141555 21.92470
#> Merc 450SL          3.730 17.350247 3.106365 11.006204 23.69429
#> Merc 450SLC         3.780 17.083024 3.109069 10.733457 23.43259
#> Cadillac Fleetwood  5.250  9.226650 3.295296  2.496758 15.95654
#> Lincoln Continental 5.424  8.296712 3.330101  1.495740 15.09769
#> Chrysler Imperial   5.345  8.718926 3.313990  1.950855 15.48700
#> Fiat 128            2.200 25.527289 3.144962 19.104419 31.95016
#> Honda Civic         1.615 28.653805 3.220219 22.077240 35.23037
#> Toyota Corolla      1.835 27.478021 3.188191 20.966866 33.98918
#> Toyota Corona       2.465 24.111004 3.121571 17.735905 30.48610
#> Dodge Challenger    3.520 18.472586 3.097736 12.146165 24.79901
#> AMC Javelin         3.435 18.926866 3.095503 12.605006 25.24873
#> Camaro Z28          3.840 16.762355 3.112643 10.405490 23.11922
#> Pontiac Firebird    3.845 16.735633 3.112957 10.378127 23.09314
#> Fiat X1-9           1.935 26.943574 3.175102 20.459150 33.42800
#> Porsche 914-2       2.140 25.847957 3.151202 19.412345 32.28357
#> Lotus Europa        1.513 29.198941 3.236547 22.589030 35.80885
#> Ford Pantera L      3.170 20.343151 3.093221 14.025952 26.66035
#> Ferrari Dino        2.770 22.480940 3.103199 16.143362 28.81852
#> Maserati Bora       3.570 18.205363 3.099389 11.875565 24.53516
#> Volvo 142E          2.780 22.427495 3.102754 16.090827 28.76416

estimate_relation(model)

#>          wt Predicted        SE    CI_low  CI_high
#> 1  1.513000 29.198941 1.0944578 26.963760 31.43412
#> 2  1.947556 26.876471 0.8909876 25.056831 28.69611
#> 3  2.382111 24.554001 0.7126988 23.098476 26.00953
#> 4  2.816667 22.231531 0.5831635 21.040552 23.42251
#> 5  3.251222 19.909061 0.5387759 18.808734 21.00939
#> 6  3.685778 17.586592 0.5987811 16.363718 18.80947
#> 7  4.120333 15.264122 0.7381448 13.756629 16.77161
#> 8  4.554889 12.941652 0.9215401 11.059616 14.82369
#> 9  4.989444 10.619182 1.1276853  8.316141 12.92222
#> 10 5.424000  8.296712 1.3461693  5.547468 11.04596

# Logistic Models
model <- glm(vs ~ wt, data = mtcars, family = "binomial")
estimate_response(model)

#>                        wt  Predicted       CI_low      CI_high
#> Mazda RX4           2.620 0.67019044 2.220446e-16 1.000000e+00
#> Mazda RX4 Wag       2.875 0.55523783 2.220446e-16 1.000000e+00
#> Datsun 710          2.320 0.78282588 2.220446e-16 1.000000e+00
#> Hornet 4 Drive      3.215 0.39466867 2.220446e-16 1.000000e+00
#> Hornet Sportabout   3.440 0.29784048 2.220446e-16 1.000000e+00
#> Valiant             3.460 0.28991157 2.220446e-16 1.000000e+00
#> Duster 360          3.570 0.24862226 2.220446e-16 1.000000e+00
#> Merc 240D           3.190 0.40613518 2.220446e-16 1.000000e+00
#> Merc 230            3.150 0.42469141 2.220446e-16 1.000000e+00
#> Merc 280            3.440 0.29784048 2.220446e-16 1.000000e+00
#> Merc 280C           3.440 0.29784048 2.220446e-16 1.000000e+00
#> Merc 450SE          4.070 0.11292079 2.220446e-16 1.000000e+00
#> Merc 450SL          3.730 0.19597214 2.220446e-16 1.000000e+00
#> Merc 450SLC         3.780 0.18135578 2.220446e-16 1.000000e+00
#> Cadillac Fleetwood  5.250 0.01318121 2.220446e-16 2.220446e-16
#> Lincoln Continental 5.424 0.00948861 2.220446e-16 2.220446e-16
#> Chrysler Imperial   5.345 0.01101745 2.220446e-16 2.220446e-16
#> Fiat 128            2.200 0.81927993 2.220446e-16 1.000000e+00
#> Honda Civic         1.615 0.93271404 2.220446e-16 1.000000e+00
#> Toyota Corolla      1.835 0.90103934 2.220446e-16 1.000000e+00
#> Toyota Corona       2.465 0.73207609 2.220446e-16 1.000000e+00
#> Dodge Challenger    3.520 0.26689250 2.220446e-16 1.000000e+00
#> AMC Javelin         3.435 0.29984210 2.220446e-16 1.000000e+00
#> Camaro Z28          3.840 0.16495379 2.220446e-16 1.000000e+00
#> Pontiac Firebird    3.845 0.16364217 2.220446e-16 1.000000e+00
#> Fiat X1-9           1.935 0.88265040 2.220446e-16 1.000000e+00
#> Porsche 914-2       2.140 0.83563597 2.220446e-16 1.000000e+00
#> Lotus Europa        1.513 0.94396025 2.220446e-16 1.000000e+00
#> Ford Pantera L      3.170 0.41538330 2.220446e-16 1.000000e+00
#> Ferrari Dino        2.770 0.60407280 2.220446e-16 1.000000e+00
#> Maserati Bora       3.570 0.24862226 2.220446e-16 1.000000e+00
#> Volvo 142E          2.780 0.59949440 2.220446e-16 1.000000e+00

estimate_relation(model)

#>          wt  Predicted         SE       CI_low   CI_high
#> 1  1.513000 0.94396025 0.06549629 0.5980458895 0.9947836
#> 2  1.947556 0.88014286 0.10010589 0.5334198865 0.9792388
#> 3  2.382111 0.76197619 0.12409535 0.4557422335 0.9244626
#> 4  2.816667 0.58256343 0.11811766 0.3500822761 0.7833487
#> 5  3.251222 0.37826154 0.10617035 0.2007286531 0.5957721
#> 6  3.685778 0.20962741 0.10102636 0.0743117139 0.4670288
#> 7  4.120333 0.10364064 0.07991040 0.0209726481 0.3842652
#> 8  4.554889 0.04798689 0.05225091 0.0053284850 0.3217030
#> 9  4.989444 0.02150163 0.03031776 0.0013024456 0.2702071
#> 10 5.424000 0.00948861 0.01640034 0.0003132392 0.2265267

# Mixed models
if (require("lme4")) {
  model <- lmer(mpg ~ wt + (1 | gear), data = mtcars)
  estimate_response(model)
  estimate_relation(model)
}

#>          wt Predicted        SE    CI_low  CI_high
#> 1  1.513000 28.556963 1.3710323 25.748531 31.36540
#> 2  1.947556 26.364627 1.1782480 23.951095 28.77816
#> 3  2.382111 24.172291 1.0253547 22.071947 26.27263
#> 4  2.816667 21.979954 0.9321916 20.070446 23.88946
#> 5  3.251222 19.787618 0.9171450 17.908932 21.66630
#> 6  3.685778 17.595282 0.9838059 15.580047 19.61052
#> 7  4.120333 15.402945 1.1176484 13.113546 17.69234
#> 8  4.554889 13.210609 1.2980551 10.551664 15.86955
#> 9  4.989444 11.018273 1.5084101  7.928435 14.10811
#> 10 5.424000  8.825937 1.7378723  5.266066 12.38581

# Bayesian models
if (require("rstanarm")) {
  model <- rstanarm::stan_glm(mpg ~ wt, data = mtcars, refresh=0)
  estimate_response(model)
  estimate_relation(model)
}

#>          wt Predicted        SE    CI_low  CI_high
#> 1  1.513000 29.180253 1.1716877 26.865488 31.47500
#> 2  1.947556 26.859637 0.9516804 24.928473 28.63410
#> 3  2.382111 24.539020 0.7522710 23.040706 25.95555
#> 4  2.816667 22.218403 0.6087395 21.003294 23.39196
#> 5  3.251222 19.897786 0.5645489 18.760959 20.97093
#> 6  3.685778 17.577170 0.6354078 16.274467 18.75940
#> 7  4.120333 15.256553 0.7923116 13.752388 16.81048
#> 8  4.554889 12.935936 0.9952561 11.065049 14.90769
#> 9  4.989444 10.615320 1.2002690  8.225173 12.96357
#> 10 5.424000  8.294703 1.4346843  5.626438 11.30207