Generates predictions for Frequentist models

Generates predictions for Frequentist models

# S3 method for glm
estimate_response(
  model,
  data = NULL,
  transform = "response",
  random = FALSE,
  length = 25,
  preserve_range = TRUE,
  predict = "response",
  ci = 0.95,
  ...
)

# S3 method for glm
estimate_link(
  model,
  data = "grid",
  transform = "response",
  random = FALSE,
  length = 25,
  preserve_range = TRUE,
  predict = "link",
  ci = 0.95,
  ...
)

Arguments

model
A statistical model.

data
A data frame with model's predictors to estimate the response. If NULL, the model's data is used. If "grid", the model matrix is obtained (through visualisation_matrix).

transform
Can be "none" (default for contrasts), "response" (default for means), "mu", "unlink", "log". "none" will leave the values on scale of the linear predictors. "response" will transform them on scale of the response variable. Thus for a logistic model, "none" will give estimations expressed in log-odds (probabilities on logit scale) and "response" in terms of probabilities.

random
Should it take the random effects into account? Can be TRUE, FALSE or a formula indicating which group-level parameters to condition on when making predictions. The data argument may include new levels of the grouping factors that were specified when the model was estimated, in which case the resulting posterior predictions marginalize over the relevant variables (see posterior_predict.stanreg).

length
Passed to visualisation_matrix if data = "grid".

preserve_range
Passed to visualisation_matrix if data = "grid".

predict
Can be "response" (default) or "link". The former predicts the the outcome per se, while the latter predicts the link function (i.e., the regression "line"), equivalent to estimating the fit. In other words, estimate_response(model, predict="link") is equivalent to estimate_link(model).

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

...
Arguments passed to or from other methods.

Value

A dataframe of predicted values.

Examples

library(modelbased)

model <- lm(Sepal.Width ~ Species * Petal.Length, data = iris)
estimate_response(model)

#>        Species Petal.Length Predicted   CI_low  CI_high
#> 1       setosa          1.4  3.403952 3.309971 3.497933
#> 2       setosa          1.4  3.403952 3.309971 3.497933
#> 3       setosa          1.3  3.365164 3.243690 3.486639
#> 4       setosa          1.5  3.442739 3.352198 3.533281
#> 5       setosa          1.4  3.403952 3.309971 3.497933
#> 6       setosa          1.7  3.520314 3.369334 3.671294
#> 7       setosa          1.4  3.403952 3.309971 3.497933
#> 8       setosa          1.5  3.442739 3.352198 3.533281
#> 9       setosa          1.4  3.403952 3.309971 3.497933
#> 10      setosa          1.5  3.442739 3.352198 3.533281
#> 11      setosa          1.5  3.442739 3.352198 3.533281
#> 12      setosa          1.6  3.481527 3.368160 3.594893
#> 13      setosa          1.4  3.403952 3.309971 3.497933
#> 14      setosa          1.1  3.287590 3.081508 3.493671
#> 15      setosa          1.2  3.326377 3.165230 3.487524
#> 16      setosa          1.5  3.442739 3.352198 3.533281
#> 17      setosa          1.3  3.365164 3.243690 3.486639
#> 18      setosa          1.4  3.403952 3.309971 3.497933
#> 19      setosa          1.7  3.520314 3.369334 3.671294
#> 20      setosa          1.5  3.442739 3.352198 3.533281
#> 21      setosa          1.7  3.520314 3.369334 3.671294
#> 22      setosa          1.5  3.442739 3.352198 3.533281
#> 23      setosa          1.0  3.248802 2.995307 3.502298
#> 24      setosa          1.7  3.520314 3.369334 3.671294
#> 25      setosa          1.9  3.597889 3.355922 3.839855
#> 26      setosa          1.6  3.481527 3.368160 3.594893
#> 27      setosa          1.6  3.481527 3.368160 3.594893
#> 28      setosa          1.5  3.442739 3.352198 3.533281
#> 29      setosa          1.4  3.403952 3.309971 3.497933
#> 30      setosa          1.6  3.481527 3.368160 3.594893
#> 31      setosa          1.6  3.481527 3.368160 3.594893
#> 32      setosa          1.5  3.442739 3.352198 3.533281
#> 33      setosa          1.5  3.442739 3.352198 3.533281
#> 34      setosa          1.4  3.403952 3.309971 3.497933
#> 35      setosa          1.5  3.442739 3.352198 3.533281
#> 36      setosa          1.2  3.326377 3.165230 3.487524
#> 37      setosa          1.3  3.365164 3.243690 3.486639
#> 38      setosa          1.4  3.403952 3.309971 3.497933
#> 39      setosa          1.3  3.365164 3.243690 3.486639
#> 40      setosa          1.5  3.442739 3.352198 3.533281
#> 41      setosa          1.3  3.365164 3.243690 3.486639
#> 42      setosa          1.3  3.365164 3.243690 3.486639
#> 43      setosa          1.3  3.365164 3.243690 3.486639
#> 44      setosa          1.6  3.481527 3.368160 3.594893
#> 45      setosa          1.9  3.597889 3.355922 3.839855
#> 46      setosa          1.4  3.403952 3.309971 3.497933
#> 47      setosa          1.6  3.481527 3.368160 3.594893
#> 48      setosa          1.4  3.403952 3.309971 3.497933
#> 49      setosa          1.5  3.442739 3.352198 3.533281
#> 50      setosa          1.4  3.403952 3.309971 3.497933
#> 51  versicolor          4.7  2.934695 2.813006 3.056384
#> 52  versicolor          4.5  2.859834 2.760354 2.959314
#> 53  versicolor          4.9  3.009556 2.859191 3.159922
#> 54  versicolor          4.0  2.672680 2.571401 2.773959
#> 55  versicolor          4.6  2.897264 2.787760 3.006768
#> 56  versicolor          4.5  2.859834 2.760354 2.959314
#> 57  versicolor          4.7  2.934695 2.813006 3.056384
#> 58  versicolor          3.3  2.410665 2.207929 2.613402
#> 59  versicolor          4.6  2.897264 2.787760 3.006768
#> 60  versicolor          3.9  2.635250 2.523461 2.747039
#> 61  versicolor          3.5  2.485527 2.316182 2.654871
#> 62  versicolor          4.2  2.747542 2.658402 2.836682
#> 63  versicolor          4.0  2.672680 2.571401 2.773959
#> 64  versicolor          4.7  2.934695 2.813006 3.056384
#> 65  versicolor          3.6  2.522957 2.369501 2.676414
#> 66  versicolor          4.4  2.822403 2.730078 2.914727
#> 67  versicolor          4.5  2.859834 2.760354 2.959314
#> 68  versicolor          4.1  2.710111 2.616620 2.803602
#> 69  versicolor          4.5  2.859834 2.760354 2.959314
#> 70  versicolor          3.9  2.635250 2.523461 2.747039
#> 71  versicolor          4.8  2.972126 2.836671 3.107580
#> 72  versicolor          4.0  2.672680 2.571401 2.773959
#> 73  versicolor          4.9  3.009556 2.859191 3.159922
#> 74  versicolor          4.7  2.934695 2.813006 3.056384
#> 75  versicolor          4.3  2.784972 2.696238 2.873706
#> 76  versicolor          4.4  2.822403 2.730078 2.914727
#> 77  versicolor          4.8  2.972126 2.836671 3.107580
#> 78  versicolor          5.0  3.046987 2.880872 3.213102
#> 79  versicolor          4.5  2.859834 2.760354 2.959314
#> 80  versicolor          3.5  2.485527 2.316182 2.654871
#> 81  versicolor          3.8  2.597819 2.473487 2.722151
#> 82  versicolor          3.7  2.560388 2.422032 2.698744
#> 83  versicolor          3.9  2.635250 2.523461 2.747039
#> 84  versicolor          5.1  3.084418 2.901933 3.266902
#> 85  versicolor          4.5  2.859834 2.760354 2.959314
#> 86  versicolor          4.5  2.859834 2.760354 2.959314
#> 87  versicolor          4.7  2.934695 2.813006 3.056384
#> 88  versicolor          4.4  2.822403 2.730078 2.914727
#> 89  versicolor          4.1  2.710111 2.616620 2.803602
#> 90  versicolor          4.0  2.672680 2.571401 2.773959
#> 91  versicolor          4.4  2.822403 2.730078 2.914727
#> 92  versicolor          4.6  2.897264 2.787760 3.006768
#> 93  versicolor          4.0  2.672680 2.571401 2.773959
#> 94  versicolor          3.3  2.410665 2.207929 2.613402
#> 95  versicolor          4.2  2.747542 2.658402 2.836682
#> 96  versicolor          4.2  2.747542 2.658402 2.836682
#> 97  versicolor          4.2  2.747542 2.658402 2.836682
#> 98  versicolor          4.3  2.784972 2.696238 2.873706
#> 99  versicolor          3.0  2.298373 2.043115 2.553631
#> 100 versicolor          4.1  2.710111 2.616620 2.803602
#> 101  virginica          6.0  3.078988 2.964659 3.193317
#> 102  virginica          5.1  2.868075 2.753334 2.982815
#> 103  virginica          5.9  3.055553 2.950735 3.160372
#> 104  virginica          5.6  2.985249 2.896501 3.073997
#> 105  virginica          5.8  3.032118 2.935029 3.129208
#> 106  virginica          6.6  3.219597 3.028353 3.410841
#> 107  virginica          4.5  2.727466 2.535648 2.919284
#> 108  virginica          6.3  3.149292 2.999405 3.299179
#> 109  virginica          5.8  3.032118 2.935029 3.129208
#> 110  virginica          6.1  3.102423 2.977206 3.227639
#> 111  virginica          5.1  2.868075 2.753334 2.982815
#> 112  virginica          5.3  2.914944 2.817586 3.012303
#> 113  virginica          5.5  2.961814 2.873007 3.050621
#> 114  virginica          5.0  2.844640 2.718964 2.970315
#> 115  virginica          5.1  2.868075 2.753334 2.982815
#> 116  virginica          5.3  2.914944 2.817586 3.012303
#> 117  virginica          5.5  2.961814 2.873007 3.050621
#> 118  virginica          6.7  3.243032 3.037303 3.448760
#> 119  virginica          6.9  3.289901 3.054539 3.525264
#> 120  virginica          5.0  2.844640 2.718964 2.970315
#> 121  virginica          5.7  3.008684 2.917090 3.100277
#> 122  virginica          4.9  2.821205 2.683557 2.958853
#> 123  virginica          6.7  3.243032 3.037303 3.448760
#> 124  virginica          4.9  2.821205 2.683557 2.958853
#> 125  virginica          5.7  3.008684 2.917090 3.100277
#> 126  virginica          6.0  3.078988 2.964659 3.193317
#> 127  virginica          4.8  2.797770 2.647360 2.948180
#> 128  virginica          4.9  2.821205 2.683557 2.958853
#> 129  virginica          5.6  2.985249 2.896501 3.073997
#> 130  virginica          5.8  3.032118 2.935029 3.129208
#> 131  virginica          6.1  3.102423 2.977206 3.227639
#> 132  virginica          6.4  3.172727 3.009494 3.335961
#> 133  virginica          5.6  2.985249 2.896501 3.073997
#> 134  virginica          5.1  2.868075 2.753334 2.982815
#> 135  virginica          5.6  2.985249 2.896501 3.073997
#> 136  virginica          6.1  3.102423 2.977206 3.227639
#> 137  virginica          5.6  2.985249 2.896501 3.073997
#> 138  virginica          5.5  2.961814 2.873007 3.050621
#> 139  virginica          4.8  2.797770 2.647360 2.948180
#> 140  virginica          5.4  2.938379 2.846614 3.030144
#> 141  virginica          5.6  2.985249 2.896501 3.073997
#> 142  virginica          5.1  2.868075 2.753334 2.982815
#> 143  virginica          5.1  2.868075 2.753334 2.982815
#> 144  virginica          5.9  3.055553 2.950735 3.160372
#> 145  virginica          5.7  3.008684 2.917090 3.100277
#> 146  virginica          5.2  2.891509 2.786342 2.996677
#> 147  virginica          5.0  2.844640 2.718964 2.970315
#> 148  virginica          5.2  2.891509 2.786342 2.996677
#> 149  virginica          5.4  2.938379 2.846614 3.030144
#> 150  virginica          5.1  2.868075 2.753334 2.982815

estimate_link(model)

#>       Species Petal.Length Predicted   CI_low  CI_high
#> 1      setosa     1.000000  3.248802 2.574224 3.923380
#> 2      setosa     1.245833  3.344155 2.703087 3.985223
#> 3      setosa     1.491667  3.439507 2.807966 4.071048
#> 4      setosa     1.737500  3.534859 2.887802 4.181916
#> 5  versicolor     3.212500  2.377914 1.715915 3.039912
#> 6  versicolor     3.458333  2.469931 1.820452 3.119410
#> 7  versicolor     3.704167  2.561948 1.921815 3.202081
#> 8  versicolor     3.950000  2.653965 2.019865 3.288064
#> 9  versicolor     4.195833  2.745982 2.114508 3.377456
#> 10 versicolor     4.441667  2.837999 2.205699 3.470299
#> 11 versicolor     4.687500  2.930016 2.293453 3.566579
#> 12  virginica     4.687500  2.771406 2.124737 3.418075
#> 13 versicolor     4.933333  3.022033 2.377839 3.666228
#> 14  virginica     4.933333  2.829017 2.189772 3.468261
#> 15  virginica     5.179167  2.886627 2.252395 3.520860
#> 16  virginica     5.425000  2.944238 2.312547 3.575929
#> 17  virginica     5.670833  3.001848 2.370199 3.633498
#> 18  virginica     5.916667  3.059459 2.425351 3.693567
#> 19  virginica     6.162500  3.117070 2.478031 3.756108
#> 20  virginica     6.408333  3.174680 2.528296 3.821064
#> 21  virginica     6.654167  3.232291 2.576227 3.888354
#> 22  virginica     6.900000  3.289901 2.621926 3.957877

library(lme4)

model <- lmer(Sepal.Width ~ Petal.Length + (1 | Species), data = iris)
estimate_response(model)

#> Warning: CI cannot be computed for mixed models when no random effects present in data. Provide some data with random effects, or set `ci = NULL`.

#>     Petal.Length    Species Predicted CI_low CI_high
#> 1            1.4     setosa  2.392900     NA      NA
#> 2            1.4     setosa  2.392900     NA      NA
#> 3            1.3     setosa  2.364722     NA      NA
#> 4            1.5     setosa  2.421078     NA      NA
#> 5            1.4     setosa  2.392900     NA      NA
#> 6            1.7     setosa  2.477433     NA      NA
#> 7            1.4     setosa  2.392900     NA      NA
#> 8            1.5     setosa  2.421078     NA      NA
#> 9            1.4     setosa  2.392900     NA      NA
#> 10           1.5     setosa  2.421078     NA      NA
#> 11           1.5     setosa  2.421078     NA      NA
#> 12           1.6     setosa  2.449256     NA      NA
#> 13           1.4     setosa  2.392900     NA      NA
#> 14           1.1     setosa  2.308366     NA      NA
#> 15           1.2     setosa  2.336544     NA      NA
#> 16           1.5     setosa  2.421078     NA      NA
#> 17           1.3     setosa  2.364722     NA      NA
#> 18           1.4     setosa  2.392900     NA      NA
#> 19           1.7     setosa  2.477433     NA      NA
#> 20           1.5     setosa  2.421078     NA      NA
#> 21           1.7     setosa  2.477433     NA      NA
#> 22           1.5     setosa  2.421078     NA      NA
#> 23           1.0     setosa  2.280189     NA      NA
#> 24           1.7     setosa  2.477433     NA      NA
#> 25           1.9     setosa  2.533789     NA      NA
#> 26           1.6     setosa  2.449256     NA      NA
#> 27           1.6     setosa  2.449256     NA      NA
#> 28           1.5     setosa  2.421078     NA      NA
#> 29           1.4     setosa  2.392900     NA      NA
#> 30           1.6     setosa  2.449256     NA      NA
#> 31           1.6     setosa  2.449256     NA      NA
#> 32           1.5     setosa  2.421078     NA      NA
#> 33           1.5     setosa  2.421078     NA      NA
#> 34           1.4     setosa  2.392900     NA      NA
#> 35           1.5     setosa  2.421078     NA      NA
#> 36           1.2     setosa  2.336544     NA      NA
#> 37           1.3     setosa  2.364722     NA      NA
#> 38           1.4     setosa  2.392900     NA      NA
#> 39           1.3     setosa  2.364722     NA      NA
#> 40           1.5     setosa  2.421078     NA      NA
#> 41           1.3     setosa  2.364722     NA      NA
#> 42           1.3     setosa  2.364722     NA      NA
#> 43           1.3     setosa  2.364722     NA      NA
#> 44           1.6     setosa  2.449256     NA      NA
#> 45           1.9     setosa  2.533789     NA      NA
#> 46           1.4     setosa  2.392900     NA      NA
#> 47           1.6     setosa  2.449256     NA      NA
#> 48           1.4     setosa  2.392900     NA      NA
#> 49           1.5     setosa  2.421078     NA      NA
#> 50           1.4     setosa  2.392900     NA      NA
#> 51           4.7 versicolor  3.322769     NA      NA
#> 52           4.5 versicolor  3.266413     NA      NA
#> 53           4.9 versicolor  3.379124     NA      NA
#> 54           4.0 versicolor  3.125524     NA      NA
#> 55           4.6 versicolor  3.294591     NA      NA
#> 56           4.5 versicolor  3.266413     NA      NA
#> 57           4.7 versicolor  3.322769     NA      NA
#> 58           3.3 versicolor  2.928279     NA      NA
#> 59           4.6 versicolor  3.294591     NA      NA
#> 60           3.9 versicolor  3.097346     NA      NA
#> 61           3.5 versicolor  2.984635     NA      NA
#> 62           4.2 versicolor  3.181879     NA      NA
#> 63           4.0 versicolor  3.125524     NA      NA
#> 64           4.7 versicolor  3.322769     NA      NA
#> 65           3.6 versicolor  3.012812     NA      NA
#> 66           4.4 versicolor  3.238235     NA      NA
#> 67           4.5 versicolor  3.266413     NA      NA
#> 68           4.1 versicolor  3.153702     NA      NA
#> 69           4.5 versicolor  3.266413     NA      NA
#> 70           3.9 versicolor  3.097346     NA      NA
#> 71           4.8 versicolor  3.350946     NA      NA
#> 72           4.0 versicolor  3.125524     NA      NA
#> 73           4.9 versicolor  3.379124     NA      NA
#> 74           4.7 versicolor  3.322769     NA      NA
#> 75           4.3 versicolor  3.210057     NA      NA
#> 76           4.4 versicolor  3.238235     NA      NA
#> 77           4.8 versicolor  3.350946     NA      NA
#> 78           5.0 versicolor  3.407302     NA      NA
#> 79           4.5 versicolor  3.266413     NA      NA
#> 80           3.5 versicolor  2.984635     NA      NA
#> 81           3.8 versicolor  3.069168     NA      NA
#> 82           3.7 versicolor  3.040990     NA      NA
#> 83           3.9 versicolor  3.097346     NA      NA
#> 84           5.1 versicolor  3.435480     NA      NA
#> 85           4.5 versicolor  3.266413     NA      NA
#> 86           4.5 versicolor  3.266413     NA      NA
#> 87           4.7 versicolor  3.322769     NA      NA
#> 88           4.4 versicolor  3.238235     NA      NA
#> 89           4.1 versicolor  3.153702     NA      NA
#> 90           4.0 versicolor  3.125524     NA      NA
#> 91           4.4 versicolor  3.238235     NA      NA
#> 92           4.6 versicolor  3.294591     NA      NA
#> 93           4.0 versicolor  3.125524     NA      NA
#> 94           3.3 versicolor  2.928279     NA      NA
#> 95           4.2 versicolor  3.181879     NA      NA
#> 96           4.2 versicolor  3.181879     NA      NA
#> 97           4.2 versicolor  3.181879     NA      NA
#> 98           4.3 versicolor  3.210057     NA      NA
#> 99           3.0 versicolor  2.843745     NA      NA
#> 100          4.1 versicolor  3.153702     NA      NA
#> 101          6.0  virginica  3.689081     NA      NA
#> 102          5.1  virginica  3.435480     NA      NA
#> 103          5.9  virginica  3.660903     NA      NA
#> 104          5.6  virginica  3.576369     NA      NA
#> 105          5.8  virginica  3.632725     NA      NA
#> 106          6.6  virginica  3.858148     NA      NA
#> 107          4.5  virginica  3.266413     NA      NA
#> 108          6.3  virginica  3.773614     NA      NA
#> 109          5.8  virginica  3.632725     NA      NA
#> 110          6.1  virginica  3.717258     NA      NA
#> 111          5.1  virginica  3.435480     NA      NA
#> 112          5.3  virginica  3.491836     NA      NA
#> 113          5.5  virginica  3.548191     NA      NA
#> 114          5.0  virginica  3.407302     NA      NA
#> 115          5.1  virginica  3.435480     NA      NA
#> 116          5.3  virginica  3.491836     NA      NA
#> 117          5.5  virginica  3.548191     NA      NA
#> 118          6.7  virginica  3.886325     NA      NA
#> 119          6.9  virginica  3.942681     NA      NA
#> 120          5.0  virginica  3.407302     NA      NA
#> 121          5.7  virginica  3.604547     NA      NA
#> 122          4.9  virginica  3.379124     NA      NA
#> 123          6.7  virginica  3.886325     NA      NA
#> 124          4.9  virginica  3.379124     NA      NA
#> 125          5.7  virginica  3.604547     NA      NA
#> 126          6.0  virginica  3.689081     NA      NA
#> 127          4.8  virginica  3.350946     NA      NA
#> 128          4.9  virginica  3.379124     NA      NA
#> 129          5.6  virginica  3.576369     NA      NA
#> 130          5.8  virginica  3.632725     NA      NA
#> 131          6.1  virginica  3.717258     NA      NA
#> 132          6.4  virginica  3.801792     NA      NA
#> 133          5.6  virginica  3.576369     NA      NA
#> 134          5.1  virginica  3.435480     NA      NA
#> 135          5.6  virginica  3.576369     NA      NA
#> 136          6.1  virginica  3.717258     NA      NA
#> 137          5.6  virginica  3.576369     NA      NA
#> 138          5.5  virginica  3.548191     NA      NA
#> 139          4.8  virginica  3.350946     NA      NA
#> 140          5.4  virginica  3.520013     NA      NA
#> 141          5.6  virginica  3.576369     NA      NA
#> 142          5.1  virginica  3.435480     NA      NA
#> 143          5.1  virginica  3.435480     NA      NA
#> 144          5.9  virginica  3.660903     NA      NA
#> 145          5.7  virginica  3.604547     NA      NA
#> 146          5.2  virginica  3.463658     NA      NA
#> 147          5.0  virginica  3.407302     NA      NA
#> 148          5.2  virginica  3.463658     NA      NA
#> 149          5.4  virginica  3.520013     NA      NA
#> 150          5.1  virginica  3.435480     NA      NA

estimate_link(model)

#> Warning: CI cannot be computed for mixed models when no random effects present in data. Provide some data with random effects, or set `ci = NULL`.

#>    Petal.Length Predicted CI_low CI_high
#> 1      1.000000  2.280189     NA      NA
#> 2      1.245833  2.349459     NA      NA
#> 3      1.491667  2.418730     NA      NA
#> 4      1.737500  2.488000     NA      NA
#> 5      1.983333  2.557271     NA      NA
#> 6      2.229167  2.626541     NA      NA
#> 7      2.475000  2.695812     NA      NA
#> 8      2.720833  2.765082     NA      NA
#> 9      2.966667  2.834353     NA      NA
#> 10     3.212500  2.903623     NA      NA
#> 11     3.458333  2.972894     NA      NA
#> 12     3.704167  3.042164     NA      NA
#> 13     3.950000  3.111435     NA      NA
#> 14     4.195833  3.180705     NA      NA
#> 15     4.441667  3.249976     NA      NA
#> 16     4.687500  3.319246     NA      NA
#> 17     4.933333  3.388517     NA      NA
#> 18     5.179167  3.457787     NA      NA
#> 19     5.425000  3.527058     NA      NA
#> 20     5.670833  3.596328     NA      NA
#> 21     5.916667  3.665599     NA      NA
#> 22     6.162500  3.734869     NA      NA
#> 23     6.408333  3.804140     NA      NA
#> 24     6.654167  3.873411     NA      NA
#> 25     6.900000  3.942681     NA      NA