Estimate the slopes of a numeric predictor (over different factor levels)

See the documentation for your object's class:

Frequentist models

Bayesian models (stanreg and brms)

estimate_slopes(
  model,
  trend = NULL,
  levels = NULL,
  transform = "response",
  standardize = TRUE,
  standardize_robust = FALSE,
  ci = 0.95,
  ...
)

# S3 method for glmmTMB
estimate_slopes(
  model,
  trend = NULL,
  levels = NULL,
  transform = "response",
  standardize = TRUE,
  standardize_robust = FALSE,
  ci = 0.95,
  component = c("conditional", "zero_inflated", "zi"),
  ...
)

Arguments

model
A Bayesian model.

trend
A character vector indicating the name of the numeric variable for which to compute the slopes.

levels
A character vector indicating the variables over which the slope will be computed. If NULL (default), it will select all the remaining predictors.

transform
Can be "none" (default for contrasts), "response" (default for means), "mu", "unlink", "log". "none" will leave the values on scale of the linear predictors. "response" will transform them on scale of the response variable. Thus for a logistic model, "none" will give estimations expressed in log-odds (probabilities on logit scale) and "response" in terms of probabilities.

standardize
If TRUE, adds standardized differences or coefficients.

standardize_robust
Robust standardization through MAD (Median Absolute Deviation, a robust estimate of SD) instead of regular SD.

ci
Credible Interval (CI) level. Default to 0.89 (89%). See ci for further details.

...
Arguments passed to or from other methods.

component
A character vector indicating the model component for which estimation is requested. Only applies to models from glmmTMB. Use "conditional" for the count-model or "zero_inflate" or "zi" for the zero-inflation model.

Value

A data frame of slopes.