Estimate the slopes of a numeric predictor (over different factor levels)

Estimate the slopes of a numeric predictor (over different factor levels)

# S3 method for lm
estimate_slopes(
  model,
  trend = NULL,
  levels = NULL,
  transform = "response",
  standardize = TRUE,
  standardize_robust = FALSE,
  ci = 0.95,
  ...
)

Arguments

model
A Bayesian model.

trend
A character vector indicating the name of the numeric variable for which to compute the slopes.

levels
A character vector indicating the variables over which the slope will be computed. If NULL (default), it will select all the remaining predictors.

transform
Can be "none" (default for contrasts), "response" (default for means), "mu", "unlink", "log". "none" will leave the values on scale of the linear predictors. "response" will transform them on scale of the response variable. Thus for a logistic model, "none" will give estimations expressed in log-odds (probabilities on logit scale) and "response" in terms of probabilities.

standardize
If TRUE, adds standardized differences or coefficients.

standardize_robust
Robust standardization through MAD (Median Absolute Deviation, a robust estimate of SD) instead of regular SD.

ci
Credible Interval (CI) level. Default to 0.89 (89%). See ci for further details.

...
Arguments passed to or from other methods.

Examples

library(modelbased)

model <- lm(Sepal.Width ~ Species * Petal.Length, data = iris)
estimate_slopes(model)

#> No numeric variable was specified for slope estimation. Selecting `trend = "Petal.Length"`.

#> Species    | Coefficient |   SE |  df |        95% CI | Coefficient (std.)
#> --------------------------------------------------------------------------
#> setosa     |        0.39 | 0.26 | 144 | [-0.13, 0.90] |               1.57
#> versicolor |        0.37 | 0.10 | 144 | [ 0.18, 0.56] |               1.52
#> virginica  |        0.23 | 0.08 | 144 | [ 0.07, 0.40] |               0.95