Easy 'avg_predictions' and 'avg_slopes'

The get_marginalmeans() function is a wrapper to facilitate the usage of marginaleffects::avg_predictions() and marginaleffects::avg_slopes(), providing a somewhat simpler and intuitive API to find the specifications and variables of interest. It is meanly made to for the developers to facilitate the organization and debugging, and end-users should rather use the estimate_*() series of functions.

Usage

get_marginalcontrasts(
  model,
  contrast = NULL,
  by = NULL,
  transform = "none",
  method = "pairwise",
  ci = 0.95,
  ...
)

model_marginalcontrasts(
  model,
  contrast = NULL,
  by = NULL,
  transform = "none",
  method = "pairwise",
  ci = 0.95,
  ...
)

get_marginaleffects(model, trend = NULL, by = NULL, ...)

get_marginalmeans(model, by = "auto", transform = NULL, ci = 0.95, ...)

model_marginalmeans(model, by = "auto", transform = NULL, ci = 0.95, ...)

Arguments

model

A statistical model.

contrast

A character vector indicating the name of the variable(s) for which to compute the contrasts.

by

The predictor variable(s) at which to evaluate the desired effect / mean / contrasts. Other predictors of the model that are not included here will be collapsed and "averaged" over (the effect will be estimated across them).

transform

Can be used to easily modulate the type argument in marginaleffects::avg_predictions(). Can be "none" or "response". "none" will leave the values on scale of the linear predictors. "response" will transform them on scale of the response variable. Thus for a logistic model, "none" will give estimations expressed in log-odds (probabilities on logit scale) and "response" in terms of probabilities.

method

Contrast method.

ci

Level for confidence intervals.

...

Other arguments passed, for instance, to insight::get_datagrid() or marginaleffects::avg_predictions().

trend

A character indicating the name of the variable for which to compute the slopes.

Examples

model <- lm(Sepal.Width ~ Species * Petal.Length, data = iris)

get_marginaleffects(model, trend = "Petal.Length", by = "Species")
#> 
#>          Term    Species Estimate Std. Error    z Pr(>|z|)    S   2.5 % 97.5 %
#>  Petal.Length setosa        0.388     0.2602 1.49  0.13606  2.9 -0.1221  0.898
#>  Petal.Length versicolor    0.374     0.0963 3.89  < 0.001 13.3  0.1856  0.563
#>  Petal.Length virginica     0.234     0.0819 2.86  0.00421  7.9  0.0739  0.395
#> 
#> Type:  response 
#> Columns: term, Species, estimate, std.error, statistic, p.value, s.value, conf.low, conf.high 
#> 
get_marginaleffects(model, trend = "Petal.Length", by = "Petal.Length")
#> 
#>          Term Petal.Length Estimate Std. Error    z Pr(>|z|)   S  2.5 % 97.5 %
#>  Petal.Length         1.00    0.388       0.26 1.49    0.136 2.9 -0.122  0.898
#>  Petal.Length         1.66    0.388       0.26 1.49    0.136 2.9 -0.122  0.898
#>  Petal.Length         2.31    0.388       0.26 1.49    0.136 2.9 -0.122  0.898
#>  Petal.Length         2.97    0.388       0.26 1.49    0.136 2.9 -0.122  0.898
#>  Petal.Length         3.62    0.388       0.26 1.49    0.136 2.9 -0.122  0.898
#>  Petal.Length         4.28    0.388       0.26 1.49    0.136 2.9 -0.122  0.898
#>  Petal.Length         4.93    0.388       0.26 1.49    0.136 2.9 -0.122  0.898
#>  Petal.Length         5.59    0.388       0.26 1.49    0.136 2.9 -0.122  0.898
#>  Petal.Length         6.24    0.388       0.26 1.49    0.136 2.9 -0.122  0.898
#>  Petal.Length         6.90    0.388       0.26 1.49    0.136 2.9 -0.122  0.898
#> 
#> Type:  response 
#> Columns: term, Petal.Length, estimate, std.error, statistic, p.value, s.value, conf.low, conf.high 
#> 
get_marginaleffects(model, trend = "Petal.Length", by = c("Species", "Petal.Length"))
#> 
#>          Term    Species Petal.Length Estimate Std. Error    z Pr(>|z|)    S
#>  Petal.Length setosa             1.00    0.388     0.2602 1.49  0.13601  2.9
#>  Petal.Length setosa             1.66    0.388     0.2601 1.49  0.13596  2.9
#>  Petal.Length versicolor         3.62    0.374     0.0963 3.89  < 0.001 13.3
#>  Petal.Length versicolor         4.28    0.374     0.0963 3.89  < 0.001 13.3
#>  Petal.Length versicolor         4.93    0.374     0.0963 3.89  < 0.001 13.3
#>  Petal.Length virginica          4.93    0.234     0.0819 2.86  0.00420  7.9
#>  Petal.Length virginica          5.59    0.234     0.0819 2.86  0.00421  7.9
#>  Petal.Length virginica          6.24    0.234     0.0818 2.86  0.00418  7.9
#>  Petal.Length virginica          6.90    0.234     0.0819 2.86  0.00420  7.9
#>    2.5 % 97.5 %
#>  -0.1221  0.898
#>  -0.1220  0.898
#>   0.1856  0.563
#>   0.1856  0.563
#>   0.1856  0.563
#>   0.0739  0.395
#>   0.0739  0.395
#>   0.0740  0.395
#>   0.0739  0.395
#> 
#> Type:  response 
#> Columns: term, Species, Petal.Length, estimate, std.error, statistic, p.value, s.value, conf.low, conf.high 
#> 
model <- lm(Sepal.Length ~ Species + Petal.Width, data = iris)

# By default, 'by' is set to "Species"
get_marginalmeans(model)
#> We selected `by = c("Species")`.
#> 
#>     Species Estimate Std. Error    t Pr(>|t|)     S 2.5 % 97.5 %  Df
#>  setosa         5.88     0.1969 29.9   <0.001 210.3  5.49   6.27 146
#>  versicolor     5.82     0.0723 80.5   <0.001 405.6  5.68   5.96 146
#>  virginica      5.83     0.1741 33.5   <0.001 231.4  5.49   6.17 146
#>  Petal.Width
#>          1.2
#>          1.2
#>          1.2
#> 
#> Type:  response 
#> Columns: rowid, Species, estimate, std.error, statistic, p.value, s.value, conf.low, conf.high, df, Petal.Width, Sepal.Length 
#> 

# Overall mean (close to 'mean(iris$Sepal.Length)')
get_marginalmeans(model, by = NULL)
#> 
#>  Estimate Std. Error    t Pr(>|t|)     S 2.5 % 97.5 %  Df    Species
#>      4.87     0.0737 66.1   <0.001 365.4  4.73   5.02 146 setosa    
#>      5.12     0.0719 71.1   <0.001 380.3  4.97   5.26 146 setosa    
#>      5.36     0.1013 52.9   <0.001 320.3  5.16   5.56 146 setosa    
#>      5.61     0.1439 39.0   <0.001 260.1  5.32   5.89 146 setosa    
#>      5.85     0.1910 30.6   <0.001 215.0  5.47   6.23 146 setosa    
#>      6.09     0.2400 25.4   <0.001 181.7  5.62   6.57 146 setosa    
#>      6.34     0.2900 21.9   <0.001 156.7  5.77   6.91 146 setosa    
#>      6.58     0.3404 19.3   <0.001 137.5  5.91   7.26 146 setosa    
#>      6.83     0.3912 17.5   <0.001 122.3  6.06   7.60 146 setosa    
#>      7.07     0.4422 16.0   <0.001 110.2  6.20   7.95 146 setosa    
#>      4.81     0.2472 19.5   <0.001 138.4  4.32   5.30 146 versicolor
#>      5.06     0.1980 25.5   <0.001 182.7  4.67   5.45 146 versicolor
#>      5.30     0.1505 35.2   <0.001 240.9  5.00   5.60 146 versicolor
#>      5.55     0.1070 51.8   <0.001 316.2  5.33   5.76 146 versicolor
#>      5.79     0.0747 77.5   <0.001 397.9  5.64   5.94 146 versicolor
#>      6.03     0.0711 84.8   <0.001 416.5  5.89   6.18 146 versicolor
#>      6.28     0.0994 63.2   <0.001 356.0  6.08   6.48 146 versicolor
#>      6.52     0.1416 46.1   <0.001 292.8  6.24   6.80 146 versicolor
#>      6.77     0.1886 35.9   <0.001 244.4  6.40   7.14 146 versicolor
#>      7.01     0.2375 29.5   <0.001 208.3  6.54   7.48 146 versicolor
#>      4.82     0.3795 12.7   <0.001  81.9  4.07   5.57 146 virginica 
#>      5.07     0.3288 15.4   <0.001 105.3  4.42   5.72 146 virginica 
#>      5.31     0.2784 19.1   <0.001 135.4  4.76   5.86 146 virginica 
#>      5.56     0.2286 24.3   <0.001 174.2  5.10   6.01 146 virginica 
#>      5.80     0.1799 32.2   <0.001 224.3  5.44   6.16 146 virginica 
#>      6.04     0.1335 45.3   <0.001 289.3  5.78   6.31 146 virginica 
#>      6.29     0.0929 67.7   <0.001 370.3  6.11   6.47 146 virginica 
#>      6.53     0.0690 94.7   <0.001 439.3  6.40   6.67 146 virginica 
#>      6.78     0.0790 85.8   <0.001 418.8  6.62   6.93 146 virginica 
#>      7.02     0.1143 61.4   <0.001 350.4  6.80   7.25 146 virginica 
#>  Petal.Width
#>        0.100
#>        0.367
#>        0.633
#>        0.900
#>        1.167
#>        1.433
#>        1.700
#>        1.967
#>        2.233
#>        2.500
#>        0.100
#>        0.367
#>        0.633
#>        0.900
#>        1.167
#>        1.433
#>        1.700
#>        1.967
#>        2.233
#>        2.500
#>        0.100
#>        0.367
#>        0.633
#>        0.900
#>        1.167
#>        1.433
#>        1.700
#>        1.967
#>        2.233
#>        2.500
#> 
#> Type:  response 
#> Columns: rowid, Species, Petal.Width, estimate, std.error, statistic, p.value, s.value, conf.low, conf.high, df, Sepal.Length 
#> 

# One can estimate marginal means at several values of a 'modulate' variable
get_marginalmeans(model, by = "Petal.Width", length = 3)
#> 
#>  Petal.Width Estimate Std. Error     t Pr(>|t|)     S 2.5 % 97.5 %  Df
#>          0.1     4.84     0.2167  22.3   <0.001 160.0  4.41   5.26 146
#>          1.3     5.94     0.0439 135.3   <0.001 513.6  5.85   6.02 146
#>          2.5     7.04     0.2552  27.6   <0.001 196.1  6.53   7.54 146
#> 
#> Type:  response 
#> Columns: Petal.Width, estimate, std.error, statistic, p.value, s.value, conf.low, conf.high, df 
#> 

# Interactions
model <- lm(Sepal.Width ~ Species * Petal.Length, data = iris)

get_marginalmeans(model)
#> We selected `by = c("Species")`.
#> 
#>     Species Estimate Std. Error     t Pr(>|t|)     S 2.5 % 97.5 %  Df
#>  setosa         4.32     0.5990  7.21   <0.001  35.0  3.13   5.50 144
#>  versicolor     2.58     0.0658 39.24   <0.001 259.3  2.45   2.71 144
#>  virginica      2.55     0.1535 16.63   <0.001 115.0  2.25   2.86 144
#>  Petal.Length
#>          3.76
#>          3.76
#>          3.76
#> 
#> Type:  response 
#> Columns: rowid, Species, estimate, std.error, statistic, p.value, s.value, conf.low, conf.high, df, Petal.Length, Sepal.Width 
#> 
get_marginalmeans(model, by = c("Species", "Petal.Length"), length = 2)
#> 
#>     Species Petal.Length Estimate Std. Error     t Pr(>|t|)     S 2.5 % 97.5 %
#>  setosa              1.0     3.25      0.128 25.33   <0.001 180.0 2.995   3.50
#>  setosa              6.9     5.54      1.415  3.91   <0.001  12.8 2.739   8.34
#>  versicolor          1.0     1.55      0.317  4.89   <0.001  18.5 0.924   2.18
#>  versicolor          6.9     3.76      0.258 14.58   <0.001  97.7 3.249   4.27
#>  virginica           1.0     1.91      0.375  5.08   <0.001  19.7 1.165   2.65
#>  virginica           6.9     3.29      0.119 27.63   <0.001 195.0 3.055   3.53
#>   Df
#>  144
#>  144
#>  144
#>  144
#>  144
#>  144
#> 
#> Type:  response 
#> Columns: rowid, Species, Petal.Length, estimate, std.error, statistic, p.value, s.value, conf.low, conf.high, df, Sepal.Width 
#> 
get_marginalmeans(model, by = c("Species", "Petal.Length = c(1, 3, 5)"), length = 2)
#> 
#>     Species Estimate Std. Error     t Pr(>|t|)     S 2.5 % 97.5 %  Df
#>  setosa         3.25     0.1282 25.33   <0.001 180.0 2.995   3.50 144
#>  setosa         4.02     0.4026 10.00   <0.001  58.0 3.229   4.82 144
#>  setosa         4.80     0.9216  5.21   <0.001  20.6 2.979   6.62 144
#>  versicolor     1.55     0.3166  4.89   <0.001  18.5 0.924   2.18 144
#>  versicolor     2.30     0.1291 17.80   <0.001 124.5 2.043   2.55 144
#>  versicolor     3.05     0.0840 36.26   <0.001 244.3 2.881   3.21 144
#>  virginica      1.91     0.3753  5.08   <0.001  19.7 1.165   2.65 144
#>  virginica      2.38     0.2137 11.12   <0.001  67.8 1.954   2.80 144
#>  virginica      2.84     0.0636 44.74   <0.001 284.5 2.719   2.97 144
#>  Petal.Length
#>             1
#>             3
#>             5
#>             1
#>             3
#>             5
#>             1
#>             3
#>             5
#> 
#> Type:  response 
#> Columns: rowid, Species, Petal.Length, estimate, std.error, statistic, p.value, s.value, conf.low, conf.high, df, Sepal.Width 
#>