Easy 'emmeans' and 'emtrends'

The model_emmeans function is a wrapper to facilitate the usage of emmeans::emmeans() and emmeans::emtrends(), providing a somewhat simpler and intuitive API to find the specifications and variables of interest. It is meanly made to for the developers to facilitate the organization and debugging, and end-users should rather use the estimate_* series of functions.

Usage

model_emcontrasts(
  model,
  contrast = NULL,
  at = NULL,
  fixed = NULL,
  transform = "none",
  ...
)

model_emmeans(
  model,
  at = "auto",
  fixed = NULL,
  transform = "response",
  levels = NULL,
  modulate = NULL,
  ...
)

model_emtrends(
  model,
  trend = NULL,
  at = NULL,
  fixed = NULL,
  levels = NULL,
  modulate = NULL,
  ...
)

Arguments

model

A statistical model.

contrast

A character vector indicating the name of the variable(s) for which to compute the contrasts.

at

The predictor variable(s) at which to evaluate the desired effect / mean / contrasts. Other predictors of the model that are not included here will be collapsed and "averaged" over (the effect will be estimated across them).

fixed

A character vector indicating the names of the predictors to be "fixed" (i.e., maintained), so that the estimation is made at these values.

transform

Is passed to the type argument in emmeans::emmeans(). See this vignette. Can be "none" (default for contrasts), "response" (default for means), "mu", "unlink", "log". "none" will leave the values on scale of the linear predictors. "response" will transform them on scale of the response variable. Thus for a logistic model, "none" will give estimations expressed in log-odds (probabilities on logit scale) and "response" in terms of probabilities.

...

Other arguments passed for instance to insight::get_datagrid().

levels, modulate

Deprecated, use at instead.

trend

A character indicating the name of the variable for which to compute the slopes.

Examples

# Basic usage
model <- lm(Sepal.Width ~ Species, data = iris)
model_emcontrasts(model)
#> No variable was specified for contrast estimation. Selecting `contrast = "Species"`.
#>  contrast               estimate     SE  df t.ratio p.value
#>  setosa - versicolor       0.658 0.0679 147   9.685  <.0001
#>  setosa - virginica        0.454 0.0679 147   6.683  <.0001
#>  versicolor - virginica   -0.204 0.0679 147  -3.003  0.0088
#> 
#> P value adjustment: tukey method for comparing a family of 3 estimates 

# Dealing with interactions
model <- lm(Sepal.Width ~ Species * Petal.Width, data = iris)
# By default: selects first factor
model_emcontrasts(model)
#> No variable was specified for contrast estimation. Selecting `contrast = "Species"`.
#> NOTE: Results may be misleading due to involvement in interactions
#>  contrast               estimate    SE  df t.ratio p.value
#>  setosa - versicolor       1.590 0.394 144   4.039  0.0003
#>  setosa - virginica        1.774 0.413 144   4.293  0.0001
#>  versicolor - virginica    0.184 0.145 144   1.272  0.4131
#> 
#> P value adjustment: tukey method for comparing a family of 3 estimates 
# Can also run contrasts between points of numeric
model_emcontrasts(model, contrast = "Petal.Width", length = 3)
#> NOTE: Results may be misleading due to involvement in interactions
#>  contrast  estimate    SE  df t.ratio p.value
#>  0.1 - 1.3    -1.01 0.195 144  -5.180  <.0001
#>  0.1 - 2.5    -2.02 0.390 144  -5.180  <.0001
#>  1.3 - 2.5    -1.01 0.195 144  -5.180  <.0001
#> 
#> Results are averaged over the levels of: Species 
#> P value adjustment: tukey method for comparing a family of 3 estimates 
# Or both
model_emcontrasts(model, contrast = c("Species", "Petal.Width"), length = 2)
#>  contrast                        estimate    SE  df t.ratio p.value
#>  setosa 0.1 - versicolor 0.1       1.8275 0.279 144   6.550  <.0001
#>  setosa 0.1 - virginica 0.1        1.5479 0.312 144   4.955  <.0001
#>  setosa 0.1 - setosa 2.5          -2.0093 0.977 144  -2.057  0.3158
#>  setosa 0.1 - versicolor 2.5      -0.7012 0.268 144  -2.614  0.1005
#>  setosa 0.1 - virginica 2.5        0.0325 0.112 144   0.289  0.9997
#>  versicolor 0.1 - virginica 0.1   -0.2797 0.406 144  -0.689  0.9829
#>  versicolor 0.1 - setosa 2.5      -3.8368 0.957 144  -4.009  0.0013
#>  versicolor 0.1 - versicolor 2.5  -2.5288 0.521 144  -4.858  <.0001
#>  versicolor 0.1 - virginica 2.5   -1.7951 0.282 144  -6.355  <.0001
#>  virginica 0.1 - setosa 2.5       -3.5571 0.967 144  -3.678  0.0044
#>  virginica 0.1 - versicolor 2.5   -2.2491 0.399 144  -5.642  <.0001
#>  virginica 0.1 - virginica 2.5    -1.5154 0.375 144  -4.043  0.0012
#>  setosa 2.5 - versicolor 2.5       1.3080 0.954 144   1.371  0.7441
#>  setosa 2.5 - virginica 2.5        2.0417 0.922 144   2.214  0.2379
#>  versicolor 2.5 - virginica 2.5    0.7337 0.272 144   2.699  0.0817
#> 
#> P value adjustment: tukey method for comparing a family of 6 estimates 
# Or with custom specifications
estimate_contrasts(model, contrast = c("Species", "Petal.Width=c(1, 2)"))
#> Marginal Contrasts Analysis
#> 
#> Level1       |       Level2 | Difference |         95% CI |   SE | t(144) |      p
#> ----------------------------------------------------------------------------------
#> setosa 1     |     setosa 2 |      -0.84 | [-2.05,  0.38] | 0.41 |  -2.06 | 0.316 
#> setosa 1     | versicolor 1 |       1.63 | [ 0.68,  2.59] | 0.32 |   5.09 | < .001
#> setosa 1     | versicolor 2 |       0.58 | [-0.45,  1.61] | 0.35 |   1.68 | 0.549 
#> setosa 1     |  virginica 1 |       1.73 | [ 0.68,  2.78] | 0.35 |   4.93 | < .001
#> setosa 1     |  virginica 2 |       1.10 | [ 0.17,  2.04] | 0.31 |   3.52 | 0.007 
#> setosa 2     | versicolor 2 |       1.42 | [-0.77,  3.60] | 0.73 |   1.94 | 0.384 
#> setosa 2     |  virginica 2 |       1.94 | [-0.20,  4.08] | 0.72 |   2.71 | 0.080 
#> versicolor 1 |     setosa 2 |      -2.47 | [-4.62, -0.32] | 0.72 |  -3.43 | 0.010 
#> versicolor 1 | versicolor 2 |      -1.05 | [-1.70, -0.41] | 0.22 |  -4.86 | < .001
#> versicolor 1 |  virginica 1 |       0.10 | [-0.45,  0.65] | 0.19 |   0.54 | 0.994 
#> versicolor 1 |  virginica 2 |      -0.53 | [-0.81, -0.25] | 0.09 |  -5.72 | < .001
#> versicolor 2 |  virginica 2 |       0.52 | [ 0.05,  0.99] | 0.16 |   3.31 | 0.015 
#> virginica 1  |     setosa 2 |      -2.57 | [-4.76, -0.38] | 0.73 |  -3.50 | 0.008 
#> virginica 1  | versicolor 2 |      -1.15 | [-1.83, -0.48] | 0.23 |  -5.13 | < .001
#> virginica 1  |  virginica 2 |      -0.63 | [-1.10, -0.17] | 0.16 |  -4.04 | 0.001 
#> 
#> Marginal contrasts estimated at Species, Petal.Width
#> p-value adjustment method: Holm (1979)
# Can fixate the numeric at a specific value
model_emcontrasts(model, fixed = "Petal.Width")
#> No variable was specified for contrast estimation. Selecting `contrast = "Species"`.
#> Petal.Width = 1.2:
#>  contrast               estimate    SE  df t.ratio p.value
#>  setosa - versicolor       1.590 0.394 144   4.039  0.0003
#>  setosa - virginica        1.774 0.413 144   4.293  0.0001
#>  versicolor - virginica    0.184 0.145 144   1.272  0.4131
#> 
#> P value adjustment: tukey method for comparing a family of 3 estimates 
# Or modulate it
model_emcontrasts(model, at = "Petal.Width", length = 4)
#> No variable was specified for contrast estimation. Selecting `contrast = "Species"`.
#> Petal.Width = 0.1:
#>  contrast               estimate    SE  df t.ratio p.value
#>  setosa - versicolor      1.8275 0.279 144   6.550  <.0001
#>  setosa - virginica       1.5479 0.312 144   4.955  <.0001
#>  versicolor - virginica  -0.2797 0.406 144  -0.689  0.7703
#> 
#> Petal.Width = 0.9:
#>  contrast               estimate    SE  df t.ratio p.value
#>  setosa - versicolor      1.6544 0.288 144   5.743  <.0001
#>  setosa - virginica       1.7125 0.325 144   5.276  <.0001
#>  versicolor - virginica   0.0581 0.208 144   0.280  0.9577
#> 
#> Petal.Width = 1.7:
#>  contrast               estimate    SE  df t.ratio p.value
#>  setosa - versicolor      1.4812 0.600 144   2.467  0.0390
#>  setosa - virginica       1.8771 0.597 144   3.144  0.0057
#>  versicolor - virginica   0.3959 0.113 144   3.502  0.0018
#> 
#> Petal.Width = 2.5:
#>  contrast               estimate    SE  df t.ratio p.value
#>  setosa - versicolor      1.3080 0.954 144   1.371  0.3587
#>  setosa - virginica       2.0417 0.922 144   2.214  0.0722
#>  versicolor - virginica   0.7337 0.272 144   2.699  0.0212
#> 
#> P value adjustment: tukey method for comparing a family of 3 estimates 
model <- lm(Sepal.Length ~ Species + Petal.Width, data = iris)

# By default, 'at' is set to "Species"
model_emmeans(model)
#> We selected `at = c("Species")`.
#>  Species    emmean     SE  df lower.CL upper.CL
#>  setosa       5.88 0.1969 146     5.49     6.27
#>  versicolor   5.82 0.0723 146     5.68     5.96
#>  virginica    5.83 0.1741 146     5.49     6.17
#> 
#> Confidence level used: 0.95 

# Overall mean (close to 'mean(iris$Sepal.Length)')
model_emmeans(model, at = NULL)
#>  1       emmean     SE  df lower.CL upper.CL
#>  overall   5.84 0.0393 146     5.77     5.92
#> 
#> Results are averaged over the levels of: Species 
#> Confidence level used: 0.95 

# One can estimate marginal means at several values of a 'modulate' variable
model_emmeans(model, at = "Petal.Width", length = 3)
#>  Petal.Width emmean     SE  df lower.CL upper.CL
#>          0.1   4.84 0.2167 146     4.41     5.26
#>          1.3   5.94 0.0439 146     5.85     6.02
#>          2.5   7.04 0.2552 146     6.53     7.54
#> 
#> Results are averaged over the levels of: Species 
#> Confidence level used: 0.95 

# Interactions
model <- lm(Sepal.Width ~ Species * Petal.Length, data = iris)

model_emmeans(model)
#> We selected `at = c("Species")`.
#> NOTE: Results may be misleading due to involvement in interactions
#>  Species    emmean     SE  df lower.CL upper.CL
#>  setosa       4.32 0.5990 144     3.13     5.50
#>  versicolor   2.58 0.0658 144     2.45     2.71
#>  virginica    2.55 0.1535 144     2.25     2.86
#> 
#> Confidence level used: 0.95 
model_emmeans(model, at = c("Species", "Petal.Length"), length = 2)
#>  Species    Petal.Length emmean    SE  df lower.CL upper.CL
#>  setosa              1.0   3.25 0.128 144    2.995     3.50
#>  versicolor          1.0   1.55 0.317 144    0.924     2.18
#>  virginica           1.0   1.91 0.375 144    1.165     2.65
#>  setosa              6.9   5.54 1.415 144    2.739     8.34
#>  versicolor          6.9   3.76 0.258 144    3.249     4.27
#>  virginica           6.9   3.29 0.119 144    3.055     3.53
#> 
#> Confidence level used: 0.95 
model_emmeans(model, at = c("Species", "Petal.Length = c(1, 3, 5)"), length = 2)
#>  Species    Petal.Length emmean     SE  df lower.CL upper.CL
#>  setosa                1   3.25 0.1282 144    2.995     3.50
#>  versicolor            1   1.55 0.3166 144    0.924     2.18
#>  virginica             1   1.91 0.3753 144    1.165     2.65
#>  setosa                3   4.02 0.4026 144    3.229     4.82
#>  versicolor            3   2.30 0.1291 144    2.043     2.55
#>  virginica             3   2.38 0.2137 144    1.954     2.80
#>  setosa                5   4.80 0.9216 144    2.979     6.62
#>  versicolor            5   3.05 0.0840 144    2.881     3.21
#>  virginica             5   2.84 0.0636 144    2.719     2.97
#> 
#> Confidence level used: 0.95 
model <- lm(Sepal.Width ~ Species * Petal.Length, data = iris)

model_emtrends(model)
#> No numeric variable was specified for slope estimation. Selecting `trend = "Petal.Length"`.
#>  1       Petal.Length.trend     SE  df lower.CL upper.CL
#>  overall              0.332 0.0964 144    0.142    0.523
#> 
#> Results are averaged over the levels of: Species 
#> Confidence level used: 0.95 
model_emtrends(model, at = "Species")
#> No numeric variable was specified for slope estimation. Selecting `trend = "Petal.Length"`.
#>  Species    Petal.Length.trend     SE  df lower.CL upper.CL
#>  setosa                  0.388 0.2602 144  -0.1264    0.902
#>  versicolor              0.374 0.0961 144   0.1843    0.564
#>  virginica               0.234 0.0819 144   0.0725    0.396
#> 
#> Confidence level used: 0.95 
model_emtrends(model, at = "Petal.Length")
#> No numeric variable was specified for slope estimation. Selecting `trend = "Petal.Length"`.
#> NOTE: Results may be misleading due to involvement in interactions
#>  Petal.Length Petal.Length.trend     SE  df lower.CL upper.CL
#>          1.00              0.332 0.0964 144    0.142    0.523
#>          1.66              0.332 0.0964 144    0.142    0.523
#>          2.31              0.332 0.0964 144    0.142    0.523
#>          2.97              0.332 0.0964 144    0.142    0.523
#>          3.62              0.332 0.0964 144    0.142    0.523
#>          4.28              0.332 0.0964 144    0.142    0.523
#>          4.93              0.332 0.0964 144    0.142    0.523
#>          5.59              0.332 0.0964 144    0.142    0.523
#>          6.24              0.332 0.0964 144    0.142    0.523
#>          6.90              0.332 0.0964 144    0.142    0.523
#> 
#> Results are averaged over the levels of: Species 
#> Confidence level used: 0.95 
model_emtrends(model, at = c("Species", "Petal.Length"))
#> No numeric variable was specified for slope estimation. Selecting `trend = "Petal.Length"`.
#>  Species    Petal.Length Petal.Length.trend     SE  df lower.CL upper.CL
#>  setosa             1.00              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         1.00              0.374 0.0961 144   0.1843    0.564
#>  virginica          1.00              0.234 0.0819 144   0.0725    0.396
#>  setosa             1.66              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         1.66              0.374 0.0961 144   0.1843    0.564
#>  virginica          1.66              0.234 0.0819 144   0.0725    0.396
#>  setosa             2.31              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         2.31              0.374 0.0961 144   0.1843    0.564
#>  virginica          2.31              0.234 0.0819 144   0.0725    0.396
#>  setosa             2.97              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         2.97              0.374 0.0961 144   0.1843    0.564
#>  virginica          2.97              0.234 0.0819 144   0.0725    0.396
#>  setosa             3.62              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         3.62              0.374 0.0961 144   0.1843    0.564
#>  virginica          3.62              0.234 0.0819 144   0.0725    0.396
#>  setosa             4.28              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         4.28              0.374 0.0961 144   0.1843    0.564
#>  virginica          4.28              0.234 0.0819 144   0.0725    0.396
#>  setosa             4.93              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         4.93              0.374 0.0961 144   0.1843    0.564
#>  virginica          4.93              0.234 0.0819 144   0.0725    0.396
#>  setosa             5.59              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         5.59              0.374 0.0961 144   0.1843    0.564
#>  virginica          5.59              0.234 0.0819 144   0.0725    0.396
#>  setosa             6.24              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         6.24              0.374 0.0961 144   0.1843    0.564
#>  virginica          6.24              0.234 0.0819 144   0.0725    0.396
#>  setosa             6.90              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         6.90              0.374 0.0961 144   0.1843    0.564
#>  virginica          6.90              0.234 0.0819 144   0.0725    0.396
#> 
#> Confidence level used: 0.95 

model <- lm(Petal.Length ~ poly(Sepal.Width, 4), data = iris)
model_emtrends(model)
#> No numeric variable was specified for slope estimation. Selecting `trend = "Sepal.Width"`.
#>  1       Sepal.Width.trend    SE  df lower.CL upper.CL
#>  overall             -2.67 0.548 145    -3.75    -1.58
#> 
#> Confidence level used: 0.95 
model_emtrends(model, at = "Sepal.Width")
#> No numeric variable was specified for slope estimation. Selecting `trend = "Sepal.Width"`.
#>  Sepal.Width Sepal.Width.trend    SE  df lower.CL upper.CL
#>         2.00             7.484 5.418 145   -3.225   18.192
#>         2.27             3.779 2.093 145   -0.358    7.916
#>         2.53             0.831 0.765 145   -0.681    2.342
#>         2.80            -1.337 0.706 145   -2.732    0.058
#>         3.07            -2.699 0.543 145   -3.772   -1.626
#>         3.33            -3.231 0.607 145   -4.430   -2.032
#>         3.60            -2.909 0.838 145   -4.564   -1.254
#>         3.87            -1.708 1.009 145   -3.702    0.287
#>         4.13             0.398 2.392 145   -4.330    5.125
#>         4.40             3.431 5.798 145   -8.028   14.890
#> 
#> Confidence level used: 0.95