Estimate Marginal Contrasts

Run a contrast analysis by estimating the differences between each level of a factor. See also other related functions such as estimate_means() and estimate_slopes().

Usage

estimate_contrasts(
  model,
  contrast = NULL,
  by = NULL,
  predict = NULL,
  ci = 0.95,
  p_adjust = "none",
  comparison = "pairwise",
  marginalize = "average",
  backend = getOption("modelbased_backend", "emmeans"),
  transform = NULL,
  verbose = TRUE,
  ...
)

Arguments

model

A statistical model.

contrast

A character vector indicating the name of the variable(s) for which to compute the contrasts.

by

The (focal) predictor variable(s) at which to evaluate the desired effect / mean / contrasts. Other predictors of the model that are not included here will be collapsed and "averaged" over (the effect will be estimated across them). The by argument is used to create a "reference grid" or "data grid" with representative values for the focal predictors. by can be a character (vector) naming the focal predictors (and optionally, representative values or levels), or a list of named elements. See details in insight::get_datagrid() to learn more about how to create data grids for predictors of interest.

predict

Is passed to the type argument in emmeans::emmeans() (when backend = "emmeans") or in marginaleffects::avg_predictions() (when backend = "marginaleffects"). For emmeans, see also this vignette. Valid options for `predict“ are:

backend = "emmeans": predict can be "response", "link", "mu", "unlink", or "log". If predict = NULL (default), the most appropriate transformation is selected (which usually is "response").

backend = "marginaleffects": predict can be "response", "link" or any valid type option supported by model's class predict() method (e.g., for zero-inflation models from package glmmTMB, you can choose predict = "zprob" or predict = "conditional" etc., see glmmTMB::predict.glmmTMB). By default, when predict = NULL, the most appropriate transformation is selected, which usually returns predictions or contrasts on the response-scale.

"link" will leave the values on scale of the linear predictors. "response" (or NULL) will transform them on scale of the response variable. Thus for a logistic model, "link" will give estimations expressed in log-odds (probabilities on logit scale) and "response" in terms of probabilities. To predict distributional parameters (called "dpar" in other packages), for instance when using complex formulae in brms models, the predict argument can take the value of the parameter you want to estimate, for instance "sigma", "kappa", etc.

ci

Confidence Interval (CI) level. Default to 0.95 (95%).

p_adjust

The p-values adjustment method for frequentist multiple comparisons. Can be one of "none" (default), "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "tukey" or "holm". See the p-value adjustment section in the emmeans::test documentation or ?stats::p.adjust.

comparison

Specify the type of contrasts or tests that should be carried out.

When backend = "emmeans", can be one of "pairwise", "poly", "consec", "eff", "del.eff", "mean_chg", "trt.vs.ctrl", "dunnett", "wtcon" and some more. See also method argument in emmeans::contrast and the ?emmeans::emmc-functions.

For backend = "marginaleffects", can be a numeric value, vector, or matrix, a string equation specifying the hypothesis to test, a string naming the comparison method, a formula, or a function. Strings, string equations and formula are probably the most common options and described below. For other options and detailed descriptions of those options, see also marginaleffects::comparisons and this website.

String: One of "pairwise", "reference", "sequential", "meandev" "meanotherdev" and the inverse options "revpairwise", "revreference", and "revsequential".

String equation: To identify parameters from the output, either specify the term name, or "b1", "b2" etc. to indicate rows, e.g.:"hp = drat", "b1 = b2", or "b1 + b2 + b3 = 0".

Formula: A formula like comparison ~ pairs | group, where the left-hand side indicates the type of comparison (difference or ratio), the right-hand side determines the pairs of estimates to compare (reference, sequential, or meandev). Optionally, comparisons can be carried out within subsets by indicating the grouping variable after a vertical bar ( |).

marginalize

Character string, indicating the type of marginalization. This dictates how the predictions are "averaged" over the non-focal predictors, i.e. those variables that are not specified in by or contrast.

"average" (default): Takes the mean value for non-focal numeric predictors and marginalizes over the factor levels of non-focal terms, which computes a kind of "weighted average" for the values at which these terms are hold constant. These predictions are a good representation of the sample, because all possible values and levels of the non-focal predictors are taken into account. It answers the question, "What is the predicted value for an 'average' observation in my data?". It refers to randomly picking a subject of your sample and the result you get on average. This approach is the one taken by default in the emmeans package.

"population": Non-focal predictors are marginalized over the observations in the sample, where the sample is replicated multiple times to produce "counterfactuals" and then takes the average of these predicted values (aggregated/grouped by the focal terms). It can be considered as extrapolation to a hypothetical target population. Counterfactual predictions are useful, insofar as the results can also be transferred to other contexts (Dickerman and Hernan, 2020). It answers the question, "What is the predicted for the 'average' observation in the broader target population?". It does not only refer to the actual data in your observed sample, but also "what would be if" we had more data, or if we had data from a different sample.

In other words, the distinction between marginalization types resides in whether the prediction are made for:

A specific "individual" from the sample (i.e., a specific combination of predictor values): this is what is obtained when using estimate_relation() and the other prediction functions.

An average individual from the sample: obtained with estimate_means(..., marginalize = "average")

The broader, hypothetical target population: obtained with estimate_means(..., marginalize = "population")

backend

Whether to use "emmeans" or "marginaleffects" as a backend. Results are usually very similar. The major difference will be found for mixed models, where backend = "marginaleffects" will also average across random effects levels, producing "marginal predictions" (instead of "conditional predictions", see Heiss 2022).

You can set a default backend via options(), e.g. use options(modelbased_backend = "emmeans") to use the emmeans package or options(modelbased_backend = "marginaleffects") to set marginaleffects as default backend.

transform

Deprecated, please use predict instead.

verbose

Use FALSE to silence messages and warnings.

...

Other arguments passed, for instance, to insight::get_datagrid(), to functions from the emmeans or marginaleffects package, or to process Bayesian models via bayestestR::describe_posterior(). Examples:

insight::get_datagrid(): Argument such as length or range can be used to control the (number of) representative values.

marginaleffects: Internally used functions are avg_predictions() for means and contrasts, and avg_slope() for slopes. Therefore, arguments for instance like vcov, transform, equivalence or slope can be passed to those functions.

emmeans: Internally used functions are emmeans() and emtrends(). Additional arguments can be passed to these functions.

Bayesian models: For Bayesian models, parameters are cleaned using describe_posterior(), thus, arguments like, for example, centrality, rope_range, or test are passed to that function.

Value

A data frame of estimated contrasts.

Details

The estimate_slopes(), estimate_means() and estimate_contrasts() functions are forming a group, as they are all based on marginal estimations (estimations based on a model). All three are built on the emmeans or marginaleffects package (depending on the backend argument), so reading its documentation (for instance emmeans::emmeans(), emmeans::emtrends() or this website) is recommended to understand the idea behind these types of procedures.

Model-based predictions is the basis for all that follows. Indeed, the first thing to understand is how models can be used to make predictions (see estimate_link()). This corresponds to the predicted response (or "outcome variable") given specific predictor values of the predictors (i.e., given a specific data configuration). This is why the concept of reference grid() is so important for direct predictions.

Marginal "means", obtained via estimate_means(), are an extension of such predictions, allowing to "average" (collapse) some of the predictors, to obtain the average response value at a specific predictors configuration. This is typically used when some of the predictors of interest are factors. Indeed, the parameters of the model will usually give you the intercept value and then the "effect" of each factor level (how different it is from the intercept). Marginal means can be used to directly give you the mean value of the response variable at all the levels of a factor. Moreover, it can also be used to control, or average over predictors, which is useful in the case of multiple predictors with or without interactions.

Marginal contrasts, obtained via estimate_contrasts(), are themselves at extension of marginal means, in that they allow to investigate the difference (i.e., the contrast) between the marginal means. This is, again, often used to get all pairwise differences between all levels of a factor. It works also for continuous predictors, for instance one could also be interested in whether the difference at two extremes of a continuous predictor is significant.

Finally, marginal effects, obtained via estimate_slopes(), are different in that their focus is not values on the response variable, but the model's parameters. The idea is to assess the effect of a predictor at a specific configuration of the other predictors. This is relevant in the case of interactions or non-linear relationships, when the effect of a predictor variable changes depending on the other predictors. Moreover, these effects can also be "averaged" over other predictors, to get for instance the "general trend" of a predictor over different factor levels.

Example: Let's imagine the following model lm(y ~ condition * x) where condition is a factor with 3 levels A, B and C and x a continuous variable (like age for example). One idea is to see how this model performs, and compare the actual response y to the one predicted by the model (using estimate_expectation()). Another idea is evaluate the average mean at each of the condition's levels (using estimate_means()), which can be useful to visualize them. Another possibility is to evaluate the difference between these levels (using estimate_contrasts()). Finally, one could also estimate the effect of x averaged over all conditions, or instead within each condition (using [estimate_slopes]).

Examples

# \dontrun{
# Basic usage
model <- lm(Sepal.Width ~ Species, data = iris)
estimate_contrasts(model)
#> No variable was specified for contrast estimation. Selecting `contrast =
#>   "Species"`.
#> Marginal Contrasts Analysis
#> 
#> Level1     |     Level2 | Difference |         95% CI |   SE | t(147) |      p
#> ------------------------------------------------------------------------------
#> setosa     | versicolor |       0.66 | [ 0.52,  0.79] | 0.07 |   9.69 | < .001
#> setosa     |  virginica |       0.45 | [ 0.32,  0.59] | 0.07 |   6.68 | < .001
#> versicolor |  virginica |      -0.20 | [-0.34, -0.07] | 0.07 |  -3.00 | 0.003 
#> 
#> Variable predicted: Sepal.Width
#> Predictors contrasted: Species
#> p-values are uncorrected.
#> 

# Dealing with interactions
model <- lm(Sepal.Width ~ Species * Petal.Width, data = iris)

# By default: selects first factor
estimate_contrasts(model)
#> No variable was specified for contrast estimation. Selecting `contrast =
#>   "Species"`.
#> NOTE: Results may be misleading due to involvement in interactions
#> Marginal Contrasts Analysis
#> 
#> Level1     |     Level2 | Difference |        95% CI |   SE | t(144) |      p
#> -----------------------------------------------------------------------------
#> setosa     | versicolor |       1.59 | [ 0.81, 2.37] | 0.39 |   4.04 | < .001
#> setosa     |  virginica |       1.77 | [ 0.96, 2.59] | 0.41 |   4.29 | < .001
#> versicolor |  virginica |       0.18 | [-0.10, 0.47] | 0.15 |   1.27 | 0.205 
#> 
#> Variable predicted: Sepal.Width
#> Predictors contrasted: Species
#> p-values are uncorrected.
#> 

# Can also run contrasts between points of numeric
estimate_contrasts(model, contrast = "Petal.Width", by = "Species", length = 4)
#> Marginal Contrasts Analysis
#> 
#> Level1         |         Level2 |    Species | Difference |         95% CI
#> --------------------------------------------------------------------------
#> Petal.Width0.1 | Petal.Width0.9 |     setosa |      -0.67 | [-1.31, -0.03]
#> Petal.Width0.1 | Petal.Width1.7 |     setosa |      -1.34 | [-2.63, -0.05]
#> Petal.Width0.1 | Petal.Width2.5 |     setosa |      -2.01 | [-3.94, -0.08]
#> Petal.Width0.9 | Petal.Width1.7 |     setosa |      -0.67 | [-1.31, -0.03]
#> Petal.Width0.9 | Petal.Width2.5 |     setosa |      -1.34 | [-2.63, -0.05]
#> Petal.Width1.7 | Petal.Width2.5 |     setosa |      -0.67 | [-1.31, -0.03]
#> Petal.Width0.1 | Petal.Width0.9 | versicolor |      -0.84 | [-1.19, -0.50]
#> Petal.Width0.1 | Petal.Width1.7 | versicolor |      -1.69 | [-2.37, -1.00]
#> Petal.Width0.1 | Petal.Width2.5 | versicolor |      -2.53 | [-3.56, -1.50]
#> Petal.Width0.9 | Petal.Width1.7 | versicolor |      -0.84 | [-1.19, -0.50]
#> Petal.Width0.9 | Petal.Width2.5 | versicolor |      -1.69 | [-2.37, -1.00]
#> Petal.Width1.7 | Petal.Width2.5 | versicolor |      -0.84 | [-1.19, -0.50]
#> Petal.Width0.1 | Petal.Width0.9 |  virginica |      -0.51 | [-0.75, -0.26]
#> Petal.Width0.1 | Petal.Width1.7 |  virginica |      -1.01 | [-1.50, -0.52]
#> Petal.Width0.1 | Petal.Width2.5 |  virginica |      -1.52 | [-2.26, -0.77]
#> Petal.Width0.9 | Petal.Width1.7 |  virginica |      -0.51 | [-0.75, -0.26]
#> Petal.Width0.9 | Petal.Width2.5 |  virginica |      -1.01 | [-1.50, -0.52]
#> Petal.Width1.7 | Petal.Width2.5 |  virginica |      -0.51 | [-0.75, -0.26]
#> 
#> Level1         |   SE | t(144) |      p
#> ---------------------------------------
#> Petal.Width0.1 | 0.33 |  -2.06 | 0.041 
#> Petal.Width0.1 | 0.65 |  -2.06 | 0.041 
#> Petal.Width0.1 | 0.98 |  -2.06 | 0.041 
#> Petal.Width0.9 | 0.33 |  -2.06 | 0.041 
#> Petal.Width0.9 | 0.65 |  -2.06 | 0.041 
#> Petal.Width1.7 | 0.33 |  -2.06 | 0.041 
#> Petal.Width0.1 | 0.17 |  -4.86 | < .001
#> Petal.Width0.1 | 0.35 |  -4.86 | < .001
#> Petal.Width0.1 | 0.52 |  -4.86 | < .001
#> Petal.Width0.9 | 0.17 |  -4.86 | < .001
#> Petal.Width0.9 | 0.35 |  -4.86 | < .001
#> Petal.Width1.7 | 0.17 |  -4.86 | < .001
#> Petal.Width0.1 | 0.12 |  -4.04 | < .001
#> Petal.Width0.1 | 0.25 |  -4.04 | < .001
#> Petal.Width0.1 | 0.37 |  -4.04 | < .001
#> Petal.Width0.9 | 0.12 |  -4.04 | < .001
#> Petal.Width0.9 | 0.25 |  -4.04 | < .001
#> Petal.Width1.7 | 0.12 |  -4.04 | < .001
#> 
#> Variable predicted: Sepal.Width
#> Predictors contrasted: Petal.Width
#> p-values are uncorrected.
#> 

# Or both
estimate_contrasts(model, contrast = c("Species", "Petal.Width"), length = 2)
#> Marginal Contrasts Analysis
#> 
#> Level1                    |                    Level2 | Difference
#> ------------------------------------------------------------------
#> setosa Petal.Width0.1     |     setosa Petal.Width2.5 |      -2.01
#> setosa Petal.Width0.1     | versicolor Petal.Width0.1 |       1.83
#> setosa Petal.Width0.1     | versicolor Petal.Width2.5 |      -0.70
#> setosa Petal.Width0.1     |  virginica Petal.Width0.1 |       1.55
#> setosa Petal.Width0.1     |  virginica Petal.Width2.5 |       0.03
#> setosa Petal.Width2.5     | versicolor Petal.Width2.5 |       1.31
#> setosa Petal.Width2.5     |  virginica Petal.Width2.5 |       2.04
#> versicolor Petal.Width0.1 |     setosa Petal.Width2.5 |      -3.84
#> versicolor Petal.Width0.1 | versicolor Petal.Width2.5 |      -2.53
#> versicolor Petal.Width0.1 |  virginica Petal.Width0.1 |      -0.28
#> versicolor Petal.Width0.1 |  virginica Petal.Width2.5 |      -1.80
#> versicolor Petal.Width2.5 |  virginica Petal.Width2.5 |       0.73
#> virginica Petal.Width0.1  |     setosa Petal.Width2.5 |      -3.56
#> virginica Petal.Width0.1  | versicolor Petal.Width2.5 |      -2.25
#> virginica Petal.Width0.1  |  virginica Petal.Width2.5 |      -1.52
#> 
#> Level1                    |         95% CI |   SE | t(144) |      p
#> -------------------------------------------------------------------
#> setosa Petal.Width0.1     | [-3.94, -0.08] | 0.98 |  -2.06 | 0.041 
#> setosa Petal.Width0.1     | [ 1.28,  2.38] | 0.28 |   6.55 | < .001
#> setosa Petal.Width0.1     | [-1.23, -0.17] | 0.27 |  -2.61 | 0.010 
#> setosa Petal.Width0.1     | [ 0.93,  2.17] | 0.31 |   4.95 | < .001
#> setosa Petal.Width0.1     | [-0.19,  0.25] | 0.11 |   0.29 | 0.773 
#> setosa Petal.Width2.5     | [-0.58,  3.19] | 0.95 |   1.37 | 0.172 
#> setosa Petal.Width2.5     | [ 0.22,  3.86] | 0.92 |   2.21 | 0.028 
#> versicolor Petal.Width0.1 | [-5.73, -1.95] | 0.96 |  -4.01 | < .001
#> versicolor Petal.Width0.1 | [-3.56, -1.50] | 0.52 |  -4.86 | < .001
#> versicolor Petal.Width0.1 | [-1.08,  0.52] | 0.41 |  -0.69 | 0.492 
#> versicolor Petal.Width0.1 | [-2.35, -1.24] | 0.28 |  -6.35 | < .001
#> versicolor Petal.Width2.5 | [ 0.20,  1.27] | 0.27 |   2.70 | 0.008 
#> virginica Petal.Width0.1  | [-5.47, -1.65] | 0.97 |  -3.68 | < .001
#> virginica Petal.Width0.1  | [-3.04, -1.46] | 0.40 |  -5.64 | < .001
#> virginica Petal.Width0.1  | [-2.26, -0.77] | 0.37 |  -4.04 | < .001
#> 
#> Variable predicted: Sepal.Width
#> Predictors contrasted: Species, Petal.Width
#> p-values are uncorrected.
#> 

# Or with custom specifications
estimate_contrasts(model, contrast = c("Species", "Petal.Width=c(1, 2)"))
#> Marginal Contrasts Analysis
#> 
#> Level1                  |                  Level2 | Difference |         95% CI
#> -------------------------------------------------------------------------------
#> setosa Petal.Width1     |     setosa Petal.Width2 |      -0.84 | [-1.64, -0.03]
#> setosa Petal.Width1     | versicolor Petal.Width1 |       1.63 | [ 1.00,  2.27]
#> setosa Petal.Width1     | versicolor Petal.Width2 |       0.58 | [-0.10,  1.26]
#> setosa Petal.Width1     |  virginica Petal.Width1 |       1.73 | [ 1.04,  2.43]
#> setosa Petal.Width1     |  virginica Petal.Width2 |       1.10 | [ 0.48,  1.72]
#> setosa Petal.Width2     | versicolor Petal.Width2 |       1.42 | [-0.03,  2.86]
#> setosa Petal.Width2     |  virginica Petal.Width2 |       1.94 | [ 0.52,  3.35]
#> versicolor Petal.Width1 |     setosa Petal.Width2 |      -2.47 | [-3.89, -1.05]
#> versicolor Petal.Width1 | versicolor Petal.Width2 |      -1.05 | [-1.48, -0.62]
#> versicolor Petal.Width1 |  virginica Petal.Width1 |       0.10 | [-0.27,  0.47]
#> versicolor Petal.Width1 |  virginica Petal.Width2 |      -0.53 | [-0.71, -0.35]
#> versicolor Petal.Width2 |  virginica Petal.Width2 |       0.52 | [ 0.21,  0.84]
#> virginica Petal.Width1  |     setosa Petal.Width2 |      -2.57 | [-4.02, -1.12]
#> virginica Petal.Width1  | versicolor Petal.Width2 |      -1.15 | [-1.60, -0.71]
#> virginica Petal.Width1  |  virginica Petal.Width2 |      -0.63 | [-0.94, -0.32]
#> 
#> Level1                  |   SE | t(144) |      p
#> ------------------------------------------------
#> setosa Petal.Width1     | 0.41 |  -2.06 | 0.041 
#> setosa Petal.Width1     | 0.32 |   5.09 | < .001
#> setosa Petal.Width1     | 0.35 |   1.68 | 0.096 
#> setosa Petal.Width1     | 0.35 |   4.93 | < .001
#> setosa Petal.Width1     | 0.31 |   3.52 | < .001
#> setosa Petal.Width2     | 0.73 |   1.94 | 0.055 
#> setosa Petal.Width2     | 0.72 |   2.71 | 0.008 
#> versicolor Petal.Width1 | 0.72 |  -3.43 | < .001
#> versicolor Petal.Width1 | 0.22 |  -4.86 | < .001
#> versicolor Petal.Width1 | 0.19 |   0.54 | 0.589 
#> versicolor Petal.Width1 | 0.09 |  -5.72 | < .001
#> versicolor Petal.Width2 | 0.16 |   3.31 | 0.001 
#> virginica Petal.Width1  | 0.73 |  -3.50 | < .001
#> virginica Petal.Width1  | 0.23 |  -5.13 | < .001
#> virginica Petal.Width1  | 0.16 |  -4.04 | < .001
#> 
#> Variable predicted: Sepal.Width
#> Predictors contrasted: Species, Petal.Width
#> p-values are uncorrected.
#> 

# Or modulate it
estimate_contrasts(model, by = "Petal.Width", length = 4)
#> No variable was specified for contrast estimation. Selecting `contrast =
#>   "Species"`.
#> Marginal Contrasts Analysis
#> 
#> Level1     |     Level2 | Petal.Width | Difference |        95% CI |   SE
#> -------------------------------------------------------------------------
#> setosa     | versicolor |        0.10 |       1.83 | [ 1.28, 2.38] | 0.28
#> setosa     |  virginica |        0.10 |       1.55 | [ 0.93, 2.17] | 0.31
#> versicolor |  virginica |        0.10 |      -0.28 | [-1.08, 0.52] | 0.41
#> setosa     | versicolor |        0.90 |       1.65 | [ 1.08, 2.22] | 0.29
#> setosa     |  virginica |        0.90 |       1.71 | [ 1.07, 2.35] | 0.32
#> versicolor |  virginica |        0.90 |       0.06 | [-0.35, 0.47] | 0.21
#> setosa     | versicolor |        1.70 |       1.48 | [ 0.29, 2.67] | 0.60
#> setosa     |  virginica |        1.70 |       1.88 | [ 0.70, 3.06] | 0.60
#> versicolor |  virginica |        1.70 |       0.40 | [ 0.17, 0.62] | 0.11
#> setosa     | versicolor |        2.50 |       1.31 | [-0.58, 3.19] | 0.95
#> setosa     |  virginica |        2.50 |       2.04 | [ 0.22, 3.86] | 0.92
#> versicolor |  virginica |        2.50 |       0.73 | [ 0.20, 1.27] | 0.27
#> 
#> Level1     | t(144) |      p
#> ----------------------------
#> setosa     |   6.55 | < .001
#> setosa     |   4.95 | < .001
#> versicolor |  -0.69 | 0.492 
#> setosa     |   5.74 | < .001
#> setosa     |   5.28 | < .001
#> versicolor |   0.28 | 0.780 
#> setosa     |   2.47 | 0.015 
#> setosa     |   3.14 | 0.002 
#> versicolor |   3.50 | < .001
#> setosa     |   1.37 | 0.172 
#> setosa     |   2.21 | 0.028 
#> versicolor |   2.70 | 0.008 
#> 
#> Variable predicted: Sepal.Width
#> Predictors contrasted: Species
#> p-values are uncorrected.
#> 

# Standardized differences
estimated <- estimate_contrasts(lm(Sepal.Width ~ Species, data = iris))
#> No variable was specified for contrast estimation. Selecting `contrast =
#>   "Species"`.
standardize(estimated)
#> Marginal Contrasts Analysis (standardized)
#> 
#> Level1     |     Level2 | Difference |         95% CI |   SE | t(147) |      p
#> ------------------------------------------------------------------------------
#> setosa     | versicolor |       1.51 | [ 1.20,  1.82] | 0.16 |   9.69 | < .001
#> setosa     |  virginica |       1.04 | [ 0.73,  1.35] | 0.16 |   6.68 | < .001
#> versicolor |  virginica |      -0.47 | [-0.78, -0.16] | 0.16 |  -3.00 | 0.003 
#> 
#> Variable predicted: Sepal.Width
#> Predictors contrasted: Species
#> p-values are uncorrected.
#> 

# Other models (mixed, Bayesian, ...)
data <- iris
data$Petal.Length_factor <- ifelse(data$Petal.Length < 4.2, "A", "B")

model <- lme4::lmer(Sepal.Width ~ Species + (1 | Petal.Length_factor), data = data)
estimate_contrasts(model)
#> No variable was specified for contrast estimation. Selecting `contrast =
#>   "Species"`.
#> Marginal Contrasts Analysis
#> 
#> Level1     |     Level2 | Difference |        95% CI |   SE |   df |     t |     p
#> ----------------------------------------------------------------------------------
#> setosa     | versicolor |       0.87 | [-0.29, 2.03] | 0.09 | 1.00 |  9.54 | 0.067
#> setosa     |  virginica |       0.80 | [-0.44, 2.03] | 0.12 | 1.11 |  6.52 | 0.081
#> versicolor |  virginica |      -0.07 | [-1.04, 0.89] | 0.08 | 1.00 | -0.97 | 0.510
#> 
#> Variable predicted: Sepal.Width
#> Predictors contrasted: Species
#> p-values are uncorrected.
#> 

data <- mtcars
data$cyl <- as.factor(data$cyl)
data$am <- as.factor(data$am)

model <- rstanarm::stan_glm(mpg ~ cyl * am, data = data, refresh = 0)
estimate_contrasts(model)
#> No variable was specified for contrast estimation. Selecting `contrast =
#>   "cyl"`.
#> NOTE: Results may be misleading due to involvement in interactions
#> Marginal Contrasts Analysis
#> 
#> Level1 | Level2 | Difference |        95% CI |     pd | % in ROPE
#> -----------------------------------------------------------------
#> cyl4   |   cyl6 |       5.68 | [2.41,  8.78] | 99.98% |        0%
#> cyl4   |   cyl8 |      10.29 | [7.06, 13.50] |   100% |        0%
#> cyl6   |   cyl8 |       4.59 | [1.12,  8.05] | 99.52% |        0%
#> 
#> Variable predicted: mpg
#> Predictors contrasted: cyl
#> 
# fix `am` at value 1
estimate_contrasts(model, by = c("cyl", "am='1'"))
#> No variable was specified for contrast estimation. Selecting `contrast =
#>   "cyl"`.
#> Marginal Contrasts Analysis
#> 
#> Level1 | Level2 | am | Difference |         95% CI |     pd | % in ROPE
#> -----------------------------------------------------------------------
#> cyl4   |   cyl6 |  1 |       7.52 | [ 3.41, 11.78] | 99.98% |        0%
#> cyl4   |   cyl8 |  1 |      12.66 | [ 7.78, 17.75] |   100% |        0%
#> cyl6   |   cyl8 |  1 |       5.14 | [-0.82, 11.02] | 95.78% |     0.66%
#> 
#> Variable predicted: mpg
#> Predictors contrasted: cyl
#> 

model <- rstanarm::stan_glm(mpg ~ cyl * wt, data = data, refresh = 0)
estimate_contrasts(model)
#> No variable was specified for contrast estimation. Selecting `contrast =
#>   "cyl"`.
#> NOTE: Results may be misleading due to involvement in interactions
#> Marginal Contrasts Analysis
#> 
#> Level1 | Level2 | Difference |        95% CI |     pd | % in ROPE
#> -----------------------------------------------------------------
#> cyl4   |   cyl6 |       2.20 | [-1.41, 5.74] | 89.05% |     2.53%
#> cyl4   |   cyl8 |       4.74 | [ 1.24, 8.38] | 99.65% |        0%
#> cyl6   |   cyl8 |       2.54 | [-0.27, 5.39] | 96.38% |     1.53%
#> 
#> Variable predicted: mpg
#> Predictors contrasted: cyl
#> 
estimate_contrasts(model, by = "wt", length = 4)
#> No variable was specified for contrast estimation. Selecting `contrast =
#>   "cyl"`.
#> Marginal Contrasts Analysis
#> 
#> Level1 | Level2 |   wt | Difference |          95% CI |     pd | % in ROPE
#> --------------------------------------------------------------------------
#> cyl4   |   cyl6 | 1.51 |       5.99 | [ -3.36, 15.33] | 89.95% |     0.71%
#> cyl4   |   cyl8 | 1.51 |       9.96 | [  4.68, 15.22] | 99.98% |        0%
#> cyl6   |   cyl8 | 1.51 |       4.07 | [ -5.93, 13.98] | 79.07% |     1.34%
#> cyl4   |   cyl6 | 2.82 |       3.10 | [ -0.23,  6.35] | 96.65% |     0.71%
#> cyl4   |   cyl8 | 2.82 |       5.94 | [  2.74,  9.26] |   100% |        0%
#> cyl6   |   cyl8 | 2.82 |       2.89 | [ -0.64,  6.44] | 94.90% |     1.11%
#> cyl4   |   cyl6 | 4.12 |       0.25 | [ -7.27,  7.86] | 53.02% |     2.39%
#> cyl4   |   cyl8 | 4.12 |       1.97 | [ -3.47,  7.59] | 76.42% |     2.08%
#> cyl6   |   cyl8 | 4.12 |       1.75 | [ -4.41,  7.68] | 72.88% |     2.24%
#> cyl4   |   cyl6 | 5.42 |      -2.58 | [-17.58, 12.40] | 63.98% |     1.11%
#> cyl4   |   cyl8 | 5.42 |      -1.96 | [-11.08,  7.57] | 68.27% |     1.50%
#> cyl6   |   cyl8 | 5.42 |       0.65 | [-13.15, 13.48] | 53.77% |     1.42%
#> 
#> Variable predicted: mpg
#> Predictors contrasted: cyl
#> 

model <- rstanarm::stan_glm(
  Sepal.Width ~ Species + Petal.Width + Petal.Length,
  data = iris,
  refresh = 0
)
estimate_contrasts(model, by = "Petal.Length", test = "bf")
#> No variable was specified for contrast estimation. Selecting `contrast =
#>   "Species"`.
#> Marginal Contrasts Analysis
#> 
#> Level1     |     Level2 | Petal.Length | Difference |       95% CI |   BF
#> -------------------------------------------------------------------------
#> setosa     | versicolor |         1.00 |       1.72 | [1.38, 2.07] | 1.00
#> setosa     |  virginica |         1.00 |       2.14 | [1.63, 2.66] | 1.00
#> versicolor |  virginica |         1.00 |       0.41 | [0.20, 0.63] | 1.00
#> setosa     | versicolor |         1.66 |       1.72 | [1.38, 2.07] | 1.00
#> setosa     |  virginica |         1.66 |       2.14 | [1.63, 2.66] | 1.00
#> versicolor |  virginica |         1.66 |       0.41 | [0.20, 0.63] | 1.00
#> setosa     | versicolor |         2.31 |       1.72 | [1.38, 2.07] | 1.00
#> setosa     |  virginica |         2.31 |       2.14 | [1.63, 2.66] | 1.00
#> versicolor |  virginica |         2.31 |       0.41 | [0.20, 0.63] | 1.00
#> setosa     | versicolor |         2.97 |       1.72 | [1.38, 2.07] | 1.00
#> setosa     |  virginica |         2.97 |       2.14 | [1.63, 2.66] | 1.00
#> versicolor |  virginica |         2.97 |       0.41 | [0.20, 0.63] | 1.00
#> setosa     | versicolor |         3.62 |       1.72 | [1.38, 2.07] | 1.00
#> setosa     |  virginica |         3.62 |       2.14 | [1.63, 2.66] | 1.00
#> versicolor |  virginica |         3.62 |       0.41 | [0.20, 0.63] | 1.00
#> setosa     | versicolor |         4.28 |       1.72 | [1.38, 2.07] | 1.00
#> setosa     |  virginica |         4.28 |       2.14 | [1.63, 2.66] | 1.00
#> versicolor |  virginica |         4.28 |       0.41 | [0.20, 0.63] | 1.00
#> setosa     | versicolor |         4.93 |       1.72 | [1.38, 2.07] | 1.00
#> setosa     |  virginica |         4.93 |       2.14 | [1.63, 2.66] | 1.00
#> versicolor |  virginica |         4.93 |       0.41 | [0.20, 0.63] | 1.00
#> setosa     | versicolor |         5.59 |       1.72 | [1.38, 2.07] | 1.00
#> setosa     |  virginica |         5.59 |       2.14 | [1.63, 2.66] | 1.00
#> versicolor |  virginica |         5.59 |       0.41 | [0.20, 0.63] | 1.00
#> setosa     | versicolor |         6.24 |       1.72 | [1.38, 2.07] | 1.00
#> setosa     |  virginica |         6.24 |       2.14 | [1.63, 2.66] | 1.00
#> versicolor |  virginica |         6.24 |       0.41 | [0.20, 0.63] | 1.00
#> setosa     | versicolor |         6.90 |       1.72 | [1.38, 2.07] | 1.00
#> setosa     |  virginica |         6.90 |       2.14 | [1.63, 2.66] | 1.00
#> versicolor |  virginica |         6.90 |       0.41 | [0.20, 0.63] | 1.00
#> 
#> Variable predicted: Sepal.Width
#> Predictors contrasted: Species
#> 
# }