Generates predictions for Frequentist models

Generates predictions for Frequentist models

# S3 method for glm
estimate_response(
  model,
  data = NULL,
  transform = "response",
  random = FALSE,
  length = 25,
  preserve_range = TRUE,
  predict = "response",
  ci = 0.95,
  ...
)

# S3 method for glm
estimate_link(
  model,
  data = "grid",
  transform = "response",
  random = FALSE,
  length = 25,
  preserve_range = TRUE,
  predict = "link",
  ci = 0.95,
  ...
)

Arguments

model
A statistical model.

data
A data frame with model's predictors to estimate the response. If NULL, the model's data is used. If "grid", the model matrix is obtained (through visualisation_matrix).

transform
Can be "none" (default for contrasts), "response" (default for means), "mu", "unlink", "log". "none" will leave the values on scale of the linear predictors. "response" will transform them on scale of the response variable. Thus for a logistic model, "none" will give estimations expressed in log-odds (probabilities on logit scale) and "response" in terms of probabilities.

random
Should it take the random effects into account? Can be TRUE, FALSE or a formula indicating which group-level parameters to condition on when making predictions. The data argument may include new levels of the grouping factors that were specified when the model was estimated, in which case the resulting posterior predictions marginalize over the relevant variables (see posterior_predict.stanreg).

length
Passed to visualisation_matrix if data = "grid".

preserve_range
Passed to visualisation_matrix if data = "grid".

predict
Can be "response" (default) or "link". The former predicts the the outcome per se, while the latter predicts the link function (i.e., the regression "line"), equivalent to estimating the fit. In other words, estimate_response(model, predict="link") is equivalent to estimate_link(model).

ci
Credible Interval (CI) level. Default to 0.89 (89%). See ci for further details.

...
Arguments passed to or from other methods.

Value

A dataframe of predicted values.

Examples

library(modelbased)

model <- lm(Sepal.Width ~ Species * Petal.Length, data = iris)
estimate_response(model)

#>        Species Petal.Length Predicted   CI_low  CI_high
#> 1       setosa          1.4  3.403952 2.771791 4.036113
#> 2       setosa          1.4  3.403952 2.771791 4.036113
#> 3       setosa          1.3  3.365164 2.728335 4.001994
#> 4       setosa          1.5  3.442739 2.811080 4.074398
#> 5       setosa          1.4  3.403952 2.771791 4.036113
#> 6       setosa          1.7  3.520314 2.877204 4.163424
#> 7       setosa          1.4  3.403952 2.771791 4.036113
#> 8       setosa          1.5  3.442739 2.811080 4.074398
#> 9       setosa          1.4  3.403952 2.771791 4.036113
#> 10      setosa          1.5  3.442739 2.811080 4.074398
#> 11      setosa          1.5  3.442739 2.811080 4.074398
#> 12      setosa          1.6  3.481527 2.846194 4.116859
#> 13      setosa          1.4  3.403952 2.771791 4.036113
#> 14      setosa          1.1  3.287590 2.629361 3.945818
#> 15      setosa          1.2  3.326377 2.680805 3.971949
#> 16      setosa          1.5  3.442739 2.811080 4.074398
#> 17      setosa          1.3  3.365164 2.728335 4.001994
#> 18      setosa          1.4  3.403952 2.771791 4.036113
#> 19      setosa          1.7  3.520314 2.877204 4.163424
#> 20      setosa          1.5  3.442739 2.811080 4.074398
#> 21      setosa          1.7  3.520314 2.877204 4.163424
#> 22      setosa          1.5  3.442739 2.811080 4.074398
#> 23      setosa          1.0  3.248802 2.574224 3.923380
#> 24      setosa          1.7  3.520314 2.877204 4.163424
#> 25      setosa          1.9  3.597889 2.927558 4.268219
#> 26      setosa          1.6  3.481527 2.846194 4.116859
#> 27      setosa          1.6  3.481527 2.846194 4.116859
#> 28      setosa          1.5  3.442739 2.811080 4.074398
#> 29      setosa          1.4  3.403952 2.771791 4.036113
#> 30      setosa          1.6  3.481527 2.846194 4.116859
#> 31      setosa          1.6  3.481527 2.846194 4.116859
#> 32      setosa          1.5  3.442739 2.811080 4.074398
#> 33      setosa          1.5  3.442739 2.811080 4.074398
#> 34      setosa          1.4  3.403952 2.771791 4.036113
#> 35      setosa          1.5  3.442739 2.811080 4.074398
#> 36      setosa          1.2  3.326377 2.680805 3.971949
#> 37      setosa          1.3  3.365164 2.728335 4.001994
#> 38      setosa          1.4  3.403952 2.771791 4.036113
#> 39      setosa          1.3  3.365164 2.728335 4.001994
#> 40      setosa          1.5  3.442739 2.811080 4.074398
#> 41      setosa          1.3  3.365164 2.728335 4.001994
#> 42      setosa          1.3  3.365164 2.728335 4.001994
#> 43      setosa          1.3  3.365164 2.728335 4.001994
#> 44      setosa          1.6  3.481527 2.846194 4.116859
#> 45      setosa          1.9  3.597889 2.927558 4.268219
#> 46      setosa          1.4  3.403952 2.771791 4.036113
#> 47      setosa          1.6  3.481527 2.846194 4.116859
#> 48      setosa          1.4  3.403952 2.771791 4.036113
#> 49      setosa          1.5  3.442739 2.811080 4.074398
#> 50      setosa          1.4  3.403952 2.771791 4.036113
#> 51  versicolor          4.7  2.934695 2.297825 3.571565
#> 52  versicolor          4.5  2.859834 2.226832 3.492836
#> 53  versicolor          4.9  3.009556 2.366590 3.652522
#> 54  versicolor          4.0  2.672680 2.039393 3.305968
#> 55  versicolor          4.6  2.897264 2.262610 3.531919
#> 56  versicolor          4.5  2.859834 2.226832 3.492836
#> 57  versicolor          4.7  2.934695 2.297825 3.571565
#> 58  versicolor          3.3  2.410665 1.753476 3.067855
#> 59  versicolor          4.6  2.897264 2.262610 3.531919
#> 60  versicolor          3.9  2.635250 2.000197 3.270302
#> 61  versicolor          3.5  2.485527 1.837859 3.133194
#> 62  versicolor          4.2  2.747542 2.116082 3.379001
#> 63  versicolor          4.0  2.672680 2.039393 3.305968
#> 64  versicolor          4.7  2.934695 2.297825 3.571565
#> 65  versicolor          3.6  2.522957 1.879262 3.166653
#> 66  versicolor          4.4  2.822403 2.190486 3.454320
#> 67  versicolor          4.5  2.859834 2.226832 3.492836
#> 68  versicolor          4.1  2.710111 2.078022 3.342199
#> 69  versicolor          4.5  2.859834 2.226832 3.492836
#> 70  versicolor          3.9  2.635250 2.000197 3.270302
#> 71  versicolor          4.8  2.972126 2.332483 3.611769
#> 72  versicolor          4.0  2.672680 2.039393 3.305968
#> 73  versicolor          4.9  3.009556 2.366590 3.652522
#> 74  versicolor          4.7  2.934695 2.297825 3.571565
#> 75  versicolor          4.3  2.784972 2.153570 3.416375
#> 76  versicolor          4.4  2.822403 2.190486 3.454320
#> 77  versicolor          4.8  2.972126 2.332483 3.611769
#> 78  versicolor          5.0  3.046987 2.400157 3.693817
#> 79  versicolor          4.5  2.859834 2.226832 3.492836
#> 80  versicolor          3.5  2.485527 1.837859 3.133194
#> 81  versicolor          3.8  2.597819 1.960438 3.235199
#> 82  versicolor          3.7  2.560388 1.920124 3.200652
#> 83  versicolor          3.9  2.635250 2.000197 3.270302
#> 84  versicolor          5.1  3.084418 2.433191 3.735644
#> 85  versicolor          4.5  2.859834 2.226832 3.492836
#> 86  versicolor          4.5  2.859834 2.226832 3.492836
#> 87  versicolor          4.7  2.934695 2.297825 3.571565
#> 88  versicolor          4.4  2.822403 2.190486 3.454320
#> 89  versicolor          4.1  2.710111 2.078022 3.342199
#> 90  versicolor          4.0  2.672680 2.039393 3.305968
#> 91  versicolor          4.4  2.822403 2.190486 3.454320
#> 92  versicolor          4.6  2.897264 2.262610 3.531919
#> 93  versicolor          4.0  2.672680 2.039393 3.305968
#> 94  versicolor          3.3  2.410665 1.753476 3.067855
#> 95  versicolor          4.2  2.747542 2.116082 3.379001
#> 96  versicolor          4.2  2.747542 2.116082 3.379001
#> 97  versicolor          4.2  2.747542 2.116082 3.379001
#> 98  versicolor          4.3  2.784972 2.153570 3.416375
#> 99  versicolor          3.0  2.298373 1.623131 2.973616
#> 100 versicolor          4.1  2.710111 2.078022 3.342199
#> 101  virginica          6.0  3.078988 2.443483 3.714493
#> 102  virginica          5.1  2.868075 2.232495 3.503654
#> 103  virginica          5.9  3.055553 2.421690 3.689416
#> 104  virginica          5.6  2.985249 2.353844 3.616653
#> 105  virginica          5.8  3.032118 2.399488 3.664749
#> 106  virginica          6.6  3.219597 2.565862 3.873332
#> 107  virginica          4.5  2.727466 2.073562 3.381369
#> 108  virginica          6.3  3.149292 2.506438 3.792147
#> 109  virginica          5.8  3.032118 2.399488 3.664749
#> 110  virginica          6.1  3.102423 2.464869 3.739976
#> 111  virginica          5.1  2.868075 2.232495 3.503654
#> 112  virginica          5.3  2.914944 2.282272 3.547616
#> 113  virginica          5.5  2.961814 2.330401 3.593227
#> 114  virginica          5.0  2.844640 2.206996 3.482284
#> 115  virginica          5.1  2.868075 2.232495 3.503654
#> 116  virginica          5.3  2.914944 2.282272 3.547616
#> 117  virginica          5.5  2.961814 2.330401 3.593227
#> 118  virginica          6.7  3.243032 2.584913 3.901150
#> 119  virginica          6.9  3.289901 2.621926 3.957877
#> 120  virginica          5.0  2.844640 2.206996 3.482284
#> 121  virginica          5.7  3.008684 2.376873 3.640494
#> 122  virginica          4.9  2.821205 2.181094 3.461316
#> 123  virginica          6.7  3.243032 2.584913 3.901150
#> 124  virginica          4.9  2.821205 2.181094 3.461316
#> 125  virginica          5.7  3.008684 2.376873 3.640494
#> 126  virginica          6.0  3.078988 2.443483 3.714493
#> 127  virginica          4.8  2.797770 2.154794 3.440747
#> 128  virginica          4.9  2.821205 2.181094 3.461316
#> 129  virginica          5.6  2.985249 2.353844 3.616653
#> 130  virginica          5.8  3.032118 2.399488 3.664749
#> 131  virginica          6.1  3.102423 2.464869 3.739976
#> 132  virginica          6.4  3.172727 2.526631 3.818824
#> 133  virginica          5.6  2.985249 2.353844 3.616653
#> 134  virginica          5.1  2.868075 2.232495 3.503654
#> 135  virginica          5.6  2.985249 2.353844 3.616653
#> 136  virginica          6.1  3.102423 2.464869 3.739976
#> 137  virginica          5.6  2.985249 2.353844 3.616653
#> 138  virginica          5.5  2.961814 2.330401 3.593227
#> 139  virginica          4.8  2.797770 2.154794 3.440747
#> 140  virginica          5.4  2.938379 2.306543 3.570215
#> 141  virginica          5.6  2.985249 2.353844 3.616653
#> 142  virginica          5.1  2.868075 2.232495 3.503654
#> 143  virginica          5.1  2.868075 2.232495 3.503654
#> 144  virginica          5.9  3.055553 2.421690 3.689416
#> 145  virginica          5.7  3.008684 2.376873 3.640494
#> 146  virginica          5.2  2.891509 2.257589 3.525430
#> 147  virginica          5.0  2.844640 2.206996 3.482284
#> 148  virginica          5.2  2.891509 2.257589 3.525430
#> 149  virginica          5.4  2.938379 2.306543 3.570215
#> 150  virginica          5.1  2.868075 2.232495 3.503654

estimate_link(model)

#>       Species Petal.Length Predicted   CI_low  CI_high
#> 1      setosa     1.000000  3.248802 2.995307 3.502298
#> 2      setosa     1.245833  3.344155 3.202123 3.486186
#> 3      setosa     1.491667  3.439507 3.349793 3.529221
#> 4      setosa     1.737500  3.534859 3.367864 3.701854
#> 5  versicolor     3.212500  2.377914 2.160091 2.595736
#> 6  versicolor     3.458333  2.469931 2.293784 2.646078
#> 7  versicolor     3.704167  2.561948 2.424200 2.699696
#> 8  versicolor     3.950000  2.653965 2.547725 2.760204
#> 9  versicolor     4.195833  2.745982 2.656737 2.835227
#> 10 versicolor     4.441667  2.837999 2.743089 2.932909
#> 11 versicolor     4.687500  2.930016 2.809947 3.050086
#> 12  virginica     4.687500  2.771406 2.605922 2.936890
#> 13 versicolor     4.933333  3.022033 2.866499 3.177568
#> 14  virginica     4.933333  2.829017 2.695458 2.962575
#> 15  virginica     5.179167  2.886627 2.779596 2.993658
#> 16  virginica     5.425000  2.944238 2.853473 3.035002
#> 17  virginica     5.670833  3.001848 2.911374 3.092323
#> 18  virginica     5.916667  3.059459 2.953167 3.165751
#> 19  virginica     6.162500  3.117070 2.984499 3.249641
#> 20  virginica     6.408333  3.174680 3.010312 3.339049
#> 21  virginica     6.654167  3.232291 3.033233 3.431348
#> 22  virginica     6.900000  3.289901 3.054539 3.525264

if (require("lme4")) {
  model <- lmer(Sepal.Width ~ Petal.Length + (1 | Species), data = iris)
  estimate_response(model)
  estimate_link(model)
}

#> Warning: CI cannot be computed for mixed models when no random effects present in data. Provide some data with random effects, or set `ci = NULL`.

#> Warning: CI cannot be computed for mixed models when no random effects present in data. Provide some data with random effects, or set `ci = NULL`.

#>    Petal.Length Predicted CI_low CI_high
#> 1      1.000000  2.280189     NA      NA
#> 2      1.245833  2.349459     NA      NA
#> 3      1.491667  2.418730     NA      NA
#> 4      1.737500  2.488000     NA      NA
#> 5      1.983333  2.557271     NA      NA
#> 6      2.229167  2.626541     NA      NA
#> 7      2.475000  2.695812     NA      NA
#> 8      2.720833  2.765082     NA      NA
#> 9      2.966667  2.834353     NA      NA
#> 10     3.212500  2.903623     NA      NA
#> 11     3.458333  2.972894     NA      NA
#> 12     3.704167  3.042164     NA      NA
#> 13     3.950000  3.111435     NA      NA
#> 14     4.195833  3.180705     NA      NA
#> 15     4.441667  3.249976     NA      NA
#> 16     4.687500  3.319246     NA      NA
#> 17     4.933333  3.388517     NA      NA
#> 18     5.179167  3.457787     NA      NA
#> 19     5.425000  3.527058     NA      NA
#> 20     5.670833  3.596328     NA      NA
#> 21     5.916667  3.665599     NA      NA
#> 22     6.162500  3.734869     NA      NA
#> 23     6.408333  3.804140     NA      NA
#> 24     6.654167  3.873411     NA      NA
#> 25     6.900000  3.942681     NA      NA