Easy Creation of 'emmeans' Objects

The model_emmeans function is a wrapper to facilitate the usage of emmeans::emmeans() and emmeans::emtrends(), providing a somewhat simpler and smart API to find the variables of interest.

model_emmeans(
  model,
  levels = NULL,
  fixed = NULL,
  modulate = NULL,
  transform = "response",
  ...
)

model_emtrends(model, trend = NULL, levels = NULL, modulate = NULL, ...)

Arguments

model
A statistical model.

levels
A character vector or formula specifying the names of the factors over which levels to estimate the means, contrasts or slopes. If NULL (default), will automatically try to find it.

fixed
A character vector indicating the names of the predictors to be "fixed" (i.e., maintained), so that the estimation is made at these values.

modulate
A character vector indicating the names of a numeric variable along which the means or the contrasts will be estimated. Other arguments from visualisation_matrix, such as length to adjust the number of data points.

transform
Is passed to the type argument in emmeans::emmeans(). See this vignette. Can be "none" (default for contrasts), "response" (default for means), "mu", "unlink", "log". "none" will leave the values on scale of the linear predictors. "response" will transform them on scale of the response variable. Thus for a logistic model, "none" will give estimations expressed in log-odds (probabilities on logit scale) and "response" in terms of probabilities.

...
Other arguments passed for instance to visualisation_matrix.

trend
A character vector indicating the name of the numeric variable for which to compute the slopes.

Value

An emmeans object.

Examples

library(modelbased)

model <- lm(Sepal.Length ~ Species + Petal.Width, data = iris)

# By default, 'levels' is set to "Species"
model_emmeans(model)
#>  Species    emmean     SE  df lower.CL upper.CL
#>  setosa       5.88 0.1969 146     5.49     6.27
#>  versicolor   5.82 0.0723 146     5.68     5.96
#>  virginica    5.83 0.1741 146     5.49     6.17
#> 
#> Confidence level used: 0.95 

# One can estimate marginal means at several values of a 'modulate' variable
model_emmeans(model, modulate = "Petal.Width", length = 3)
#> Petal.Width = 0.1:
#>  Species    emmean     SE  df lower.CL upper.CL
#>  setosa       4.87 0.0737 146     4.73     5.02
#>  versicolor   4.81 0.2472 146     4.32     5.30
#>  virginica    4.82 0.3795 146     4.07     5.57
#> 
#> Petal.Width = 1.3:
#>  Species    emmean     SE  df lower.CL upper.CL
#>  setosa       5.97 0.2154 146     5.55     6.40
#>  versicolor   5.91 0.0682 146     5.78     6.05
#>  virginica    5.92 0.1563 146     5.61     6.23
#> 
#> Petal.Width = 2.5:
#>  Species    emmean     SE  df lower.CL upper.CL
#>  setosa       7.07 0.4422 146     6.20     7.95
#>  versicolor   7.01 0.2375 146     6.54     7.48
#>  virginica    7.02 0.1143 146     6.80     7.25
#> 
#> Confidence level used: 0.95 
model <- lm(Sepal.Width ~ Species * Petal.Length, data = iris)

model_emtrends(model)
#> No numeric variable was specified for slope estimation. Selecting `trend = "Petal.Length"`.
#>  Species    Petal.Length.trend     SE  df lower.CL upper.CL
#>  setosa                  0.388 0.2602 144  -0.1264    0.902
#>  versicolor              0.374 0.0961 144   0.1843    0.564
#>  virginica               0.234 0.0819 144   0.0725    0.396
#> 
#> Confidence level used: 0.95 
model_emtrends(model, levels = "Species")
#> No numeric variable was specified for slope estimation. Selecting `trend = "Petal.Length"`.
#>  Species    Petal.Length.trend     SE  df lower.CL upper.CL
#>  setosa                  0.388 0.2602 144  -0.1264    0.902
#>  versicolor              0.374 0.0961 144   0.1843    0.564
#>  virginica               0.234 0.0819 144   0.0725    0.396
#> 
#> Confidence level used: 0.95 
model_emtrends(model, modulate = "Petal.Length")
#> No numeric variable was specified for slope estimation. Selecting `trend = "Petal.Length"`.
#> NOTE: Results may be misleading due to involvement in interactions
#>  Petal.Length Petal.Length.trend     SE  df lower.CL upper.CL
#>          1.00              0.332 0.0964 144    0.142    0.523
#>          1.66              0.332 0.0964 144    0.142    0.523
#>          2.31              0.332 0.0964 144    0.142    0.523
#>          2.97              0.332 0.0964 144    0.142    0.523
#>          3.62              0.332 0.0964 144    0.142    0.523
#>          4.28              0.332 0.0964 144    0.142    0.523
#>          4.93              0.332 0.0964 144    0.142    0.523
#>          5.59              0.332 0.0964 144    0.142    0.523
#>          6.24              0.332 0.0964 144    0.142    0.523
#>          6.90              0.332 0.0964 144    0.142    0.523
#> 
#> Results are averaged over the levels of: Species 
#> Confidence level used: 0.95 
model_emtrends(model, levels = "Species", modulate = "Petal.Length")
#> No numeric variable was specified for slope estimation. Selecting `trend = "Petal.Length"`.
#>  Species    Petal.Length Petal.Length.trend     SE  df lower.CL upper.CL
#>  setosa             1.00              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         1.00              0.374 0.0961 144   0.1843    0.564
#>  virginica          1.00              0.234 0.0819 144   0.0725    0.396
#>  setosa             1.66              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         1.66              0.374 0.0961 144   0.1843    0.564
#>  virginica          1.66              0.234 0.0819 144   0.0725    0.396
#>  setosa             2.31              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         2.31              0.374 0.0961 144   0.1843    0.564
#>  virginica          2.31              0.234 0.0819 144   0.0725    0.396
#>  setosa             2.97              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         2.97              0.374 0.0961 144   0.1843    0.564
#>  virginica          2.97              0.234 0.0819 144   0.0725    0.396
#>  setosa             3.62              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         3.62              0.374 0.0961 144   0.1843    0.564
#>  virginica          3.62              0.234 0.0819 144   0.0725    0.396
#>  setosa             4.28              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         4.28              0.374 0.0961 144   0.1843    0.564
#>  virginica          4.28              0.234 0.0819 144   0.0725    0.396
#>  setosa             4.93              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         4.93              0.374 0.0961 144   0.1843    0.564
#>  virginica          4.93              0.234 0.0819 144   0.0725    0.396
#>  setosa             5.59              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         5.59              0.374 0.0961 144   0.1843    0.564
#>  virginica          5.59              0.234 0.0819 144   0.0725    0.396
#>  setosa             6.24              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         6.24              0.374 0.0961 144   0.1843    0.564
#>  virginica          6.24              0.234 0.0819 144   0.0725    0.396
#>  setosa             6.90              0.388 0.2602 144  -0.1264    0.902
#>  versicolor         6.90              0.374 0.0961 144   0.1843    0.564
#>  virginica          6.90              0.234 0.0819 144   0.0725    0.396
#> 
#> Confidence level used: 0.95 

model <- lm(Petal.Length ~ poly(Sepal.Width, 4), data = iris)
model_emtrends(model, modulate = "Sepal.Width")
#> No numeric variable was specified for slope estimation. Selecting `trend = "Sepal.Width"`.
#>  Sepal.Width Sepal.Width.trend    SE  df lower.CL upper.CL
#>         2.00             7.484 5.418 145   -3.225   18.192
#>         2.27             3.779 2.093 145   -0.358    7.916
#>         2.53             0.831 0.765 145   -0.681    2.342
#>         2.80            -1.337 0.706 145   -2.732    0.058
#>         3.07            -2.699 0.543 145   -3.772   -1.626
#>         3.33            -3.231 0.607 145   -4.430   -2.032
#>         3.60            -2.909 0.838 145   -4.564   -1.254
#>         3.87            -1.708 1.009 145   -3.702    0.287
#>         4.13             0.398 2.392 145   -4.330    5.125
#>         4.40             3.431 5.798 145   -8.028   14.890
#> 
#> Confidence level used: 0.95