Easy Creation of 'emmeans' Marginal Means Objects

The model_emmeans function is a wrapper to facilitate the usage of emmeans::emmeans(), providing a somewhat simpler and smart API to find the variables of interest.

model_emmeans(
  model,
  levels = NULL,
  fixed = NULL,
  modulate = NULL,
  transform = "response",
  ...
)

Arguments

model
A statistical model.

levels
A character vector or formula specifying the names of the predicting factors over which to estimate means or contrasts.

fixed
A character vector indicating the names of the predictors to be "fixed" (i.e., maintained), so that the estimation is made at these values.

modulate
A character vector indicating the names of a numeric variable along which the means or the contrasts will be estimated. Other arguments from visualisation_matrix, such as length to adjust the number of data points.

transform
Is passed to the type argument in emmeans::emmeans(). See this vignette. Can be "none" (default for contrasts), "response" (default for means), "mu", "unlink", "log". "none" will leave the values on scale of the linear predictors. "response" will transform them on scale of the response variable. Thus for a logistic model, "none" will give estimations expressed in log-odds (probabilities on logit scale) and "response" in terms of probabilities.

...
Other arguments passed for instance to visualisation_matrix.

Value

An emmeans object.

Examples

library(modelbased)

model <- lm(Sepal.Length ~ Species + Petal.Width, data = iris)

# By default, 'levels' is set to "Species"
model_emmeans(model)
#>  Species    emmean     SE  df lower.CL upper.CL
#>  setosa       5.88 0.1969 146     5.49     6.27
#>  versicolor   5.82 0.0723 146     5.68     5.96
#>  virginica    5.83 0.1741 146     5.49     6.17
#> 
#> Confidence level used: 0.95 

# One can estimate marginal means at several values of a 'modulate' variable
model_emmeans(model, modulate = "Petal.Width", length = 3)
#> Petal.Width = 0.1:
#>  Species    emmean     SE  df lower.CL upper.CL
#>  setosa       4.87 0.0737 146     4.73     5.02
#>  versicolor   4.81 0.2472 146     4.32     5.30
#>  virginica    4.82 0.3795 146     4.07     5.57
#> 
#> Petal.Width = 1.3:
#>  Species    emmean     SE  df lower.CL upper.CL
#>  setosa       5.97 0.2154 146     5.55     6.40
#>  versicolor   5.91 0.0682 146     5.78     6.05
#>  virginica    5.92 0.1563 146     5.61     6.23
#> 
#> Petal.Width = 2.5:
#>  Species    emmean     SE  df lower.CL upper.CL
#>  setosa       7.07 0.4422 146     6.20     7.95
#>  versicolor   7.01 0.2375 146     6.54     7.48
#>  virginica    7.02 0.1143 146     6.80     7.25
#> 
#> Confidence level used: 0.95